% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{actionButtonStyled}
\alias{actionButtonStyled}
\title{Re-write shint actionButton to enable styles}
\usage{
actionButtonStyled(inputId, label, icon = NULL, width = NULL,
  type = "default", btn_type = "button", class = "", ...)
}
\description{
Re-write shint actionButton to enable styles
}
