% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_afni.R
\name{afni_tools}
\alias{afni_tools}
\title{Read from AFNIio.R, return an environment wrapper for all functions}
\usage{
afni_tools(debug)
}
\description{
Read from AFNIio.R, return an environment wrapper for all functions
}
