% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-data_repo.R
\name{any_subject_loaded}
\alias{any_subject_loaded}
\title{Function to check if data repository has data}
\usage{
any_subject_loaded(rave_data = getDefaultDataRepository())
}
\arguments{
\item{rave_data}{internally used}
}
\description{
Function to check if data repository has data
}
