% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{archive_subject}
\alias{archive_subject}
\title{Archive Subject into Zipped file}
\usage{
archive_subject(
  project_name,
  subject_code,
  include_cache = FALSE,
  include_fs = TRUE,
  include_raw = FALSE,
  save_to = tempdir()
)
}
\arguments{
\item{project_name}{project name}

\item{subject_code}{subject code}

\item{include_cache}{whether to include cache for faster loading. Default is 
false}

\item{include_fs}{whether to include 'FreeSurfer' and 'AFNI/SUMA' files}

\item{include_raw}{whether to include raw data}

\item{save_to}{directory to save file to}
}
\description{
Save subject data, including brain imaging files into a zipped 
file. Notice this function does not guarantee every file is in. Please always
double check what's inside.
}
