% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file_structure.R
\name{arrange_data_dir}
\alias{arrange_data_dir}
\title{Initialize data repository}
\usage{
arrange_data_dir(first_time = FALSE, reset = FALSE)
}
\arguments{
\item{first_time}{will create data repositories for you}

\item{reset}{reset to default data directory}
}
\description{
Initialize data repository
}
