% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_default_setups.R
\name{arrange_modules}
\alias{arrange_modules}
\title{Update (optional), and check validity of modules}
\usage{
arrange_modules(first_time = FALSE, reset = F, quiet = F)
}
\arguments{
\item{first_time}{check updates}

\item{reset}{same as first_time, check module updates, ignored}

\item{quiet}{no overwrite messages}
}
\description{
Update (optional), and check validity of modules
}
