% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_subject.R
\name{as.character.Subject}
\alias{as.character.Subject}
\title{Returns subject ID}
\usage{
\method{as.character}{Subject}(x, ...)
}
<<<<<<< HEAD
=======
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
>>>>>>> dev
\description{
Returns subject ID
}
