% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Tensor.R
\name{as.vector.Tensor}
\alias{as.vector.Tensor}
\title{Convert Tensors to Vector}
\usage{
\method{as.vector}{Tensor}(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{passed from or to other methods.}
}
\description{
Convert Tensors to Vector
}
