% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{as_data_uri}
\alias{as_data_uri}
\title{File or any R objects to Data URI, base64 encoded}
\usage{
as_data_uri(obj, is_file = FALSE, mime = "text/plain;charset=US-ASCII", ...)
}
\description{
File or any R objects to Data URI, base64 encoded
}
