% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_electrode.R
\name{as_printable}
\alias{as_printable}
\title{Function to make an environment printable}
\usage{
as_printable(env)
}
\arguments{
\item{env}{environment to be converted}
}
\description{
Function to make an environment printable
}
