% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Subject.R
\name{as_subject}
\alias{as_subject}
\title{Make new subject object from character}
\usage{
as_subject(subject, strict = TRUE, reference = "default")
}
\arguments{
\item{subject}{characters in format \code{"project/subject"}}

\item{strict}{logical indication whether preprocess folder is needed}

\item{reference}{what reference file the subject is using}
}
\description{
Make new subject object from character
}
