% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-cls_module.R
\name{attachDefaultDataRepository}
\alias{attachDefaultDataRepository}
\title{Attach subject data}
\usage{
attachDefaultDataRepository(
  unload = TRUE,
  rave_data = getDefaultDataRepository()
)
}
\arguments{
\item{unload}{TRUE if you want to detach}

\item{rave_data}{internally used}
}
\description{
Attach subject data
}
