% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ecog_repo.R
\name{baseline}
\alias{baseline}
\title{Baseline signals}
\usage{
baseline(el, from, to, method = "mean", unit = "\%", data_only = F,
  hybrid = TRUE, swap_file = tempfile(), mem_optimize = TRUE,
  same_dimension = unit \%in\% c("\%", "dB"), preop = NULL, op)
}
\arguments{
\item{el}{Tensor or ECoGTensor object}

\item{from}{baseline start time}

\item{to}{baseline end time}

\item{method}{mean or median, default is mean}

\item{unit}{"\%" percent signal change or "dB" decibel unit}

\item{data_only}{return array or tensor object?}

\item{hybrid}{if return tensor object, swap cache? useful for large dataset}

\item{swap_file}{by default tempfile, or you can specify path}

\item{mem_optimize}{optimize for large dataset? default is TRUE}

\item{same_dimension}{logical, true if \code{op} is element-wise operator}

\item{preop}{function before baselining}

\item{op}{function for baselining}
}
\description{
Baseline signals
}
