% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_shinirize.R
\name{beautify}
\alias{beautify}
\title{I forget the usage but it's related to converting something to strings}
\usage{
beautify(v, max_len = 20, is_vector = TRUE, level = 0)
}
\arguments{
\item{v}{object}

\item{max_len}{lax character length}

\item{is_vector}{is vector}

\item{level}{iter level}
}
\description{
I forget the usage but it's related to converting something to strings
}
