% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_shinydashboard.R
\name{box}
\alias{box}
\title{Wrapper for "box" in shinydashboard}
\usage{
box(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = 6, height = NULL,
  collapsible = FALSE, collapsed = FALSE, headerColor = "#f4f4f4")
}
\arguments{
\item{...}{passed to shinydashboard::box}
}
\description{
Wrapper for "box" in shinydashboard
}
