% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_wavelet.R
\name{bulk_wavelet}
\alias{bulk_wavelet}
\title{Bulk run wavelet (deprecated)}
\usage{
bulk_wavelet(project_name, subject_code, blocks, channels, srate,
  target_srate = 100, frequencies = seq(4, 200, by = 4),
  wave_num = c(3, 14), compress = 1, replace = F,
  save_original = F, ncores = future::availableCores() - 2,
  plan = NULL, save_dir = "cache_dir", filename = "\%d.h5",
  reference_name = "CAR", ...)
}
\arguments{
\item{project_name}{project_name}

\item{subject_code}{subject_code}

\item{blocks}{blocks}

\item{channels}{channels}

\item{srate}{srate}

\item{target_srate}{target_srate}

\item{frequencies}{frequencies}

\item{wave_num}{wave_num}

\item{compress}{compress}

\item{replace}{replace}

\item{save_original}{save_original}

\item{ncores}{ncores}

\item{plan}{plan}

\item{save_dir}{save_dir}

\item{filename}{filename}

\item{reference_name}{reference_name}

\item{...}{other args}
}
\description{
Bulk run wavelet (deprecated)
}
