% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-cls_module.R
\name{cache}
\alias{cache}
\title{Cache object}
\usage{
cache(key, val, global = FALSE, replace = FALSE, session = NULL,
  swap = FALSE, file = tempfile(), name = "data")
}
\arguments{
\item{key}{Any R object, a named list would be the best.}

\item{val}{Value to cache, if key exists, then value will not be evaluated nor saved}

\item{global}{option for shiny app, where if global, then the the cache will ignore sessions.}

\item{replace}{Force replace cache?}

\item{session}{internally used}

\item{swap}{Save to swap? usually when val is a large matrix or vector}

\item{file, name}{If you use swap=T, see \code{\link{save_h5}}}
}
\description{
Cache object
}
\examples{
\dontrun{
cache('a', 1) # returns 1
cache('a', 2) # still returns 1

# clear cache
clear_cache()
cache('a', 2) # Now returns 2

# Not run because a is cached
cache('a', 2)
cache('a', {Sys.sleep(10); 1})

# Use swap

y = cache('aa', 1:1000000, swap = T)
object.size(1:1000000)
object.size(y)
y[1:5]
}
}
\seealso{
\code{\link{clear_cache}}
}
