% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-cls_module.R
\name{cache_input}
\alias{cache_input}
\title{Cache input values}
\usage{
cache_input(inputId, val, read_only = T)
}
\arguments{
\item{inputId}{input ID}

\item{val}{value if not cached}

\item{read_only}{logical, if FALSE, replace cache}
}
\description{
Cache input values
}
