% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_subject}
\alias{cache_subject}
\title{Cache Subject Electrodes}
\usage{
cache_subject(subject_id, electrodes = NULL, use_rhdf5 = FALSE)
}
\arguments{
\item{subject_id}{Folder name for subject}

\item{electrodes}{\code{NULL} by default, meaning caching all valid electrodes. Or you can specify a
vector of integers indicating the number of electrodes to be cached (usually for broken cache files)}

\item{use_rhdf5}{\code{FALSE} by default. HDF5 is still under test. Don't change it.}
}
\description{
Create cache for subjects
}
