% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_repo.R
\name{check_data_repo}
\alias{check_data_repo}
\title{Check if default data environment has object}
\usage{
check_data_repo(var = c("subject"), any = F)
}
\arguments{
\item{var}{variable name}

\item{any}{if true, any of variable is in data repo will return true}
}
\description{
Check if default data environment has object
}
