% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check and Install RAVE Dependencies}
\usage{
check_dependencies(update_rave = TRUE)
}
\arguments{
\item{update_rave}{logical, whether to restart `RStudio` after installation}
}
\description{
Check and Install RAVE Dependencies
}
