% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_epoch}
\alias{check_epoch}
\title{Check if epoch file is valid}
\usage{
check_epoch(subject, epoch_name)
}
\arguments{
\item{subject}{subject object or string}

\item{epoch_name}{epoch name to check}
}
\description{
Check if epoch file is valid
}
