% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_subject}
\alias{check_subject}
\title{Complete validity check a RAVE subject}
\usage{
check_subject(subject, stop_on_error = FALSE)
}
\arguments{
\item{subject}{character, must be \code{"project/subject"} format}

\item{stop_on_error}{logical, whether stop when error occurs}
}
\value{
None
}
\description{
Complete validity check a RAVE subject
}
