% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_subjects2}
\alias{check_subjects2}
\title{check subject validity tools}
\usage{
check_subjects2(project_name, subject_code, quiet = FALSE)
}
\arguments{
\item{project_name}{project_name}

\item{subject_code}{subject_code}

\item{quiet}{logical}
}
\description{
check subject validity tools
}
