% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_subjects_old}
\alias{check_subjects_old}
\title{check subject validity tools (use check_subjects2)}
\usage{
check_subjects_old(
  project_name,
  subject_code,
  check = TRUE,
  folders = c("Subject Folder", "RAVE Folder", "Preprocessing Folder", "Meta Folder",
    "Channel Folder"),
  preprocess = c("Started Preprocess", "Notch Filter", "Wavelet"),
  Meta = c("Electrode File", "Time point File", "Frequency File", "Epoch File")
)
}
\arguments{
\item{project_name}{project_name}

\item{subject_code}{subject_code}

\item{check}{check is internally used}

\item{folders}{folders to check}

\item{preprocess}{preprocess to check}

\item{Meta}{Meta to check}
}
\description{
check subject validity tools (use check_subjects2)
}
