% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cache}
\usage{
clear_cache(all = FALSE, session = NULL)
}
\arguments{
\item{all}{Clear all cache? Don't turn it on in shiny app. This is for debug use.}

\item{session}{internally used}
}
\description{
Clear cache
}
\seealso{
\code{\link{cache}}
}
