% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-main_app.R
\name{close_tab}
\alias{close_tab}
\title{Close a tab in RAVE main application}
\usage{
close_tab(module_id, tabname)
}
\arguments{
\item{module_id}{character, module ID}

\item{tabname}{character, tab box title}
}
\description{
Close a tab in RAVE main application
}
