% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{color_console}
\alias{color_console}
\title{Enable RAVE color console}
\usage{
color_console(enable = T)
}
\arguments{
\item{enable}{logical, enable or not}
}
\description{
Enable RAVE color console
}
