% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{compoundInput}
\alias{compoundInput}
\title{Customized input just designed for RAVE to include multiple inputs and duplicate them}
\usage{
compoundInput(inputId, label = "", components = NULL, max_ncomp = 10,
  inital_ncomp = 1, prefix = "Group", style = "")
}
\arguments{
\item{inputId}{shiny input ID}

\item{label}{Label prefix}

\item{components}{expressions to include ordinary shiny inputs}

\item{max_ncomp}{maximum components to show}

\item{inital_ncomp}{initial counts}

\item{prefix}{fieldset legend}

\item{style}{additional css style}
}
\description{
Customized input just designed for RAVE to include multiple inputs and duplicate them
}
