% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_resample.R
\name{compress_signal}
\alias{compress_signal}
\title{Library to re-sample data
decimate_fir is more recommended}
\usage{
compress_signal(s, p, q, d = 5)
}
\arguments{
\item{s}{signal to be re-sampled}

\item{p}{target frequency}

\item{q}{Current signal frequency}

\item{d}{tolerance}
}
\description{
Down sampling a signal after low pass filtering
}
\details{
This function is a wrapper for \code{signal::resample}:
 Note that p and q do not need to be integers since this routine does not use
 a polyphase rate change algorithm, but instead uses bandlimited
 interpolation, wherein the continuous time signal is estimated by summing
 the sinc functions of the nearest neighbouring points up to distance d.

 Note that resample computes all samples up to but not including time n+1.
 If you are increasing the sample rate, this means that it will generate
 samples beyond the end of the time range of the original signal.
}
