% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mem.R
\name{copy_env}
\alias{copy_env}
\title{Copy objects from one environment to another}
\usage{
copy_env(from_env, to_env, deep = FALSE)
}
\arguments{
\item{from_env}{from}

\item{to_env}{target}

\item{deep}{change function environments?}
}
\description{
Copy objects from one environment to another
}
