% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dashboard.R
\name{dashboardControl}
\alias{dashboardControl}
\title{Shiny dashboard-adminLTE overrides}
\usage{
dashboardControl(..., disable = FALSE, collapsed = FALSE)
}
\arguments{
\item{...}{components to be added to control panel}

\item{disable}{to show it or not}

\item{collapsed}{initialize collapsed}
}
\description{
Shiny dashboard-adminLTE overrides
}
