% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shinydashboard.R
\name{dashboardHeader}
\alias{dashboardHeader}
\title{Shiny dashboard-adminLTE overrides - dashboardHeader}
\usage{
dashboardHeader(..., title = NULL, titleWidth = NULL,
  disable = FALSE, btn_text_right = "Controls", .list = NULL)
}
\arguments{
\item{...}{additional navigation buttons}

\item{title}{header title}

\item{titleWidth}{not used}

\item{disable}{hide header}

\item{btn_text_right}{control panel button name}

\item{.list}{see ...}
}
\description{
Shiny dashboard-adminLTE overrides - dashboardHeader
}
