% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shinydashboard.R
\name{dashboardPage}
\alias{dashboardPage}
\title{Shiny dashboard-adminLTE overrides - dashboardPage}
\usage{
dashboardPage(header, sidebar, control, body, title = NULL,
  skin = c("blue", "black", "purple", "green", "red", "yellow"),
  controlbar_opened = FALSE, initial_mask = NULL)
}
\arguments{
\item{header}{dashboardHeader object}

\item{sidebar}{dashboardSidebar object}

\item{control}{dashboardControl object}

\item{body}{dashboardBody object}

\item{title}{title name}

\item{skin}{theme color}

\item{controlbar_opened}{open control panel by default or not}

\item{initial_mask}{internally used}
}
\description{
Shiny dashboard-adminLTE overrides - dashboardPage
}
