% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_repository.R
\docType{data}
\name{data_repository}
\alias{data_repository}
\title{Singleton of internal class DataRepository}
\format{An object of class \code{DataRepository} (inherits from \code{R6}) of length 13.}
\usage{
data_repository$set_data(multi_electrodes)
data_repository$get_data()
data_repository$set_global_var(key = value)
data_repository$get_global_var(key, default = NULL)
data_repository$reset()
}
\description{
\code{data_repository} is a singleton instance creating an environment to store electrode data and modules.
Usually you won't be able to use this object directly
}
\details{
DO NOT use this class directly, use the singleton \code{data_repository} instead.
This documentation only provides definitions for generating data_repository.
}
\seealso{
\code{\link{DataRepository}}
}
\keyword{datasets}
