% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_repository.R
\docType{data}
\name{DataRepository}
\alias{DataRepository}
\title{DataRepository, inner-class definition for singleton \code{data_repository}.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataRepository
}
\description{
DataRepository, inner-class definition for singleton \code{data_repository}.
}
\details{
DO NOT use this class directly, use the singleton \code{data_repository} instead.
This documentation only provides definitions for generating data_repository.
}
\section{Fields}{

\describe{
\item{\code{get_data()}}{get data environment that stores electrode data.
Returns a \code{MultiElectrodes} instance}

\item{\code{set_data(multi_electrodes)}}{set data environment,
multi_electrodes must be \code{MultiElectrodes} class}

\item{\code{add_module(module)}}{add module to the repository,
module must be \code{ModuleEnv} class}

\item{\code{get_module(id)}}{get \code{ModuleEnv} instance by its id.}

\item{\code{reset()}}{remove all data imported as well as modules stored in this repository.}
}}

\seealso{
\code{\link{MultiElectrodes}}, \code{\link{ModuleEnv}}
}
\keyword{DataRepository,}
\keyword{R6Class}
\keyword{data_repository,}
