% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_debug.R
\name{debug_module}
\alias{debug_module}
\title{Function for module writers to debug modules}
\usage{
debug_module(package = package, module_id = module_id,
  reload = FALSE)
}
\arguments{
\item{m}{module object see ModuleEnvir}

\item{private_id}{session rave_id}
}
\description{
Function for module writers to debug modules
}
\details{
To use this function, you must use test.mode=TRUE, i.e. call init_app(m, test.mode=TRUE)
otherwise objects will be removed. Once imported data, close rave session and run debug_module(m)
to see variables within the module
}
