% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_resample.R
\name{decimate_fir}
\alias{decimate_fir}
\title{Decimate signal via FIR1 filter}
\usage{
decimate_fir(x, q, n = 30)
}
\arguments{
\item{x}{signal (vector)}

\item{q}{Compress rate or downsample rate}

\item{n}{nth order of FIR kernel will be used, default 30}
}
\value{
Down-sampled signal
}
\description{
Decimate signal via FIR1 filter
}
