% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-define_initialization.R
\name{define_initialization}
\alias{define_initialization}
\alias{define_initialization.default}
\alias{define_initialization.rave_module_debug}
\alias{define_initialization.rave_running}
\alias{define_initialization.rave_running_local}
\title{Defines 'RAVE' Module Initialization
Defines the global variables for the module. Called along with 
\code{\link[rave]{define_input}} to define UI initialization
actions once a subject is loaded.}
\usage{
define_initialization(expr)

\method{define_initialization}{default}(expr)

\method{define_initialization}{rave_module_debug}(expr)

\method{define_initialization}{rave_running}(expr)

\method{define_initialization}{rave_running_local}(expr)
}
\arguments{
\item{expr}{R expression to run after subject is loaded}
}
\value{
None
}
\description{
Defines 'RAVE' Module Initialization
Defines the global variables for the module. Called along with 
\code{\link[rave]{define_input}} to define UI initialization
actions once a subject is loaded.
}
\examples{

\dontrun{
# Requires to install R package beauchamplab/ravebuiltins

# Enable debug mode
ravebuiltins::dev_ravebuiltins(reload = FALSE)
# Check data
define_initialization({
  rave_checks('power')
  power <- module_tools$get_power(referenced = TRUE)
})


# Initialize global variables for modules
ravebuiltins::dev_ravebuiltins(reload = FALSE)
define_initialization({
  print(subject$info())
  time_points = preload_info$time_points
})
define_input(
  shiny::sliderInput('time_range', 'Time-Range', min=0, 
                     max=1, value = c(0,1)),
  init_args = c('min', 'max', 'value'),
  init_expr = {
    min = min(time_points)
    max = max(time_points)
    value = c(0, max)
  }
)
}
}
