% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parse.R
\name{deparse_selections}
\alias{deparse_selections}
\title{Convert integer vector to text}
\usage{
deparse_selections(nums, link = "-", concatenate = T, max_lag = 1)
}
\arguments{
\item{nums}{integer vector}

\item{link}{character to concatenate}

\item{concatenate}{strings}

\item{max_lag}{how to define consecutive}
}
\description{
Convert integer vector to text
}
