% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_devs.R
\name{detect_modules}
\alias{detect_modules}
\title{Check all packages to for new RAVE module packages}
\usage{
detect_modules(packages, as_module = TRUE)
}
\arguments{
\item{packages}{array of packages to search for, default is all packages}

\item{as_module}{logical, try to return module instances or just a list of modules}
}
\description{
Check all packages to for new RAVE module packages
}
