% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{download_sample_data}
\alias{download_sample_data}
\title{Function to download demo data to data repository}
\usage{
download_sample_data(subject = "KC", ...)
}
\arguments{
\item{subject}{demo subject}

\item{...}{other parameters passed to \code{download_subject_data}}
}
\value{
Nothing
}
\description{
Function to download demo data to data repository
}
