% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{download_subject_data}
\alias{download_subject_data}
\title{Function to download subjects from internet/local}
\usage{
download_subject_data(
  con,
  replace_if_exists = FALSE,
  override_project = NULL,
  override_subject = NULL,
  temp_dir = tempdir(),
  remove_zipfile = TRUE,
  subject_settings = NULL,
  mode = "wb",
  ...
)
}
\arguments{
\item{con}{an url or local file path}

\item{replace_if_exists}{Automatically replace current subject if subject 
files exist (default FALSE)}

\item{override_project}{if not null, project will be renamed to this value}

\item{override_subject}{if not null, subject will be renamed to this value}

\item{temp_dir}{temp directory to store downloaded zip files and extracted 
files}

\item{remove_zipfile}{clear downloaded zip files? if \code{con} is local 
file, this will be forced to FALSE}

\item{subject_settings}{override \code{"subject.yaml"} see details}

\item{mode, ...}{passed to \code{\link[utils]{download.file}}}
}
\description{
Function to download subjects from internet/local
}
\details{
Each downloaded zip file should have a \code{"subject.yaml"} file indicating 
default project name, subject code, data directory and raw data directory. 

If you want to override subject settings, you need to implement your own
\code{subject_settings}. See examples.
}
\examples{
\dontrun{
# Normal example
download.file(
  'https://s3-us-west-2.amazonaws.com/rave-demo-subject/sfn-demo/data-large.zip',
  destfile = "~/rave_data/data-small.zip", mode = "wb")
download_subject_data(con = "~/rave_data/data-small.zip")

# or the following
# download_subject_data(
# 'https://s3-us-west-2.amazonaws.com/rave-demo-subject/sfn-demo/data-large.zip'
# )

# rename project to demo_junk
download_subject_data(con = "~/rave_data/data-small.zip",
           override_project = 'demo_junk')

# override settings
download_subject_data(
  con = "~/rave_data/data-small.zip",
  subject_settings = list(
    # subject conf
    'demo_project/demo_subject' = list(
      data_dir = 'data 2/data_dir/demo/sub1',
      raw_dir = 'data 2/raw_dir/sub1'
    )
  )
)
}
}
