% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkvalid.R
\name{dropInvalid}
\alias{dropInvalid}
\title{Similar to dropNulls}
\usage{
dropInvalid(x, deep = FALSE, nulls = c(""))
}
\arguments{
\item{x}{list to drop}

\item{deep}{iterativly apply this function}

\item{nulls}{see is_invalid}
}
\description{
Similar to dropNulls
}
