% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{dropNulls}
\alias{dropNulls}
\title{Drop nulls within lists/vectors}
\usage{
dropNulls(x, .invalids = c("null"))
}
\arguments{
\item{x}{list}

\item{.invalids}{'null', 'na', 'blank' default is null}
}
\description{
Drop nulls within lists/vectors
}
\examples{
\dontrun{
x <- list(NULL,NULL,1,2)
dropNulls(x)
}
}
