% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ECoGRepository.R
\docType{data}
\name{ECoGRepository}
\alias{ECoGRepository}
\title{R6 class for ECoG data Repository}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ECoGRepository
}
\description{
R6 class for ECoG data Repository
}
\keyword{datasets}
