% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ECoGTensor.R
\name{ECoGTensor}
\alias{ECoGTensor}
\title{ECoG Tensor class inherit from Tensor}
\description{
ECoG Tensor class inherit from Tensor

ECoG Tensor class inherit from Tensor
}
\section{Super class}{
\code{\link[rave:Tensor]{rave::Tensor}} -> \code{ECoGTensor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-flatten}{\code{ECoGTensor$flatten()}}
\item \href{#method-new}{\code{ECoGTensor$new()}}
\item \href{#method-clone}{\code{ECoGTensor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id=".use_multi_files">}\href{../../rave/html/Tensor.html#method-.use_multi_files}{\code{rave::Tensor$.use_multi_files()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="collapse">}\href{../../rave/html/Tensor.html#method-collapse}{\code{rave::Tensor$collapse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="finalize">}\href{../../rave/html/Tensor.html#method-finalize}{\code{rave::Tensor$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="get_data">}\href{../../rave/html/Tensor.html#method-get_data}{\code{rave::Tensor$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="operate">}\href{../../rave/html/Tensor.html#method-operate}{\code{rave::Tensor$operate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="print">}\href{../../rave/html/Tensor.html#method-print}{\code{rave::Tensor$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="set_data">}\href{../../rave/html/Tensor.html#method-set_data}{\code{rave::Tensor$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="subset">}\href{../../rave/html/Tensor.html#method-subset}{\code{rave::Tensor$subset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="to_swap">}\href{../../rave/html/Tensor.html#method-to_swap}{\code{rave::Tensor$to_swap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="to_swap_now">}\href{../../rave/html/Tensor.html#method-to_swap_now}{\code{rave::Tensor$to_swap_now()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-flatten"></a>}}
\subsection{Method \code{flatten()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGTensor$flatten(include_index = T, value_name = "value")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGTensor$new(
  data,
  dim,
  dimnames,
  varnames,
  hybrid = F,
  swap_file = tempfile(),
  temporary = TRUE,
  multi_files = FALSE,
  use_index = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGTensor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
