% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_electrode.R
\docType{data}
\name{Electrode}
\alias{Electrode}
\title{R6 class of electrode}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Electrode
}
\description{
R6 class of electrode
}
\keyword{datasets}
