% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-localization.R
\name{electrode_localization}
\alias{electrode_localization}
\title{Electrode localization}
\usage{
electrode_localization(subject_code, freesurfer_path, ct_path, ...)
}
\arguments{
\item{subject_code}{'RAVE' subject code}

\item{freesurfer_path}{'FreeSurfer' folder path that points to reconstructed
subject brain}

\item{ct_path}{'CT' path (in 'Nifti' format). The 'CT' has to be aligned to 
'T1-MRI'. Please check \href{https://dipterix.org/threeBrain/articles/B-electrode-localization.html#ct-co-registration}{this tutorial}.}

\item{...}{other parameters passing to \code{\link[threeBrain]{localization_module}}}
}
\value{
This function will launch a shiny application.
}
\description{
Electrode localization
}
