% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{eval_dirty}
\alias{eval_dirty}
\title{Evaluate expressions}
\usage{
eval_dirty(expr, env = parent,frame(), data = NULL)
}
\description{
Evaluate expressions
}
\details{
\code{eval_dirty} uses \code{base::eval()} function to evaluate expressions.
Compare to \code{rlang::eval_tidy}, which won't affect original environment,
\code{eval_dirty} will cause changes to the environment. Therefore if \code{expr}
contains assignment, environment will be changed in this case.
}
\examples{
\dontrun{
expr = quote(a <- 111)
a = 1; env = globalenv()
rlang::eval_tidy(expr, env)
print(a)  # Will be 1
eval_dirty(expr, env)
print(a)  # a is changed
}
}
