% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-eval_when_ready.R
\name{eval_when_ready}
\alias{eval_when_ready}
\title{Add Function to run once Module is Ready}
\usage{
eval_when_ready(FUN)
}
\arguments{
\item{FUN}{function that takes an environment (runtime environment) as
arguments.}
}
\description{
Usually contains reactive functions that requires shiny reactive
context
}
