% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{export_diagnose_voltage}
\alias{export_diagnose_voltage}
\title{Export voltage diagnostics for electrodes}
\usage{
export_diagnose_voltage(subject, electrodes, blocks,
  save_dir = "./export", width = 12, height = 7, useDingbats = F,
  onefile = T, winlen, freq_lim, nclass = 50, fore_col = "black",
  back_col = "grey80", ...)
}
\arguments{
\item{subject}{"project_name/subject_code"}

\item{electrodes}{for example: c(1,3,4,10:15) will generate diagnose plots for 9 electrodes}

\item{blocks}{default is all blocks of this subject}

\item{save_dir}{which directory you want to export default will create a folder "exportr" in current path}

\item{width, height, useDingbats}{passed to \code{\link{pdf}}}

\item{onefile}{collect files within one file?}

\item{winlen}{window length, default is 2*subject_sample_rate}

\item{freq_lim}{default is half of voltage sample rate, however I woould recommend: 300 Hz}

\item{nclass}{number of classes in histogram plot, default is 50}

\item{fore_col}{Notch filtered color}

\item{back_col}{Raw filtered color}

\item{...}{All other params will be passed to \code{\link{pdf}} function}
}
\description{
You must import subject throught rave_preprocess() and then run this function
}
\details{
Run "rave_preprocess()" first, import subject. If the subject is notch filtered, then it will compare
both raw and filtered signals, otherwise it will just show the raw voltage plots
}
\seealso{
\code{\link{pdf}}, \code{\link{diagnose_signal}}, \code{\link{pwelch}}
}
