% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{fake_session}
\alias{fake_session}
\title{internally used to cheat RAVE}
\usage{
fake_session(rave_id = "__fake_session__")
}
\arguments{
\item{rave_id}{internally used}
}
\description{
internally used to cheat RAVE
}
