% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{fileInputMinimal}
\alias{fileInputMinimal}
\title{Re-write shiny `fileInput`, but minimal}
\usage{
fileInputMinimal(inputId, label, multiple = FALSE, accept = NULL,
  width = "auto", type = "default", class = "", ...)
}
\arguments{
\item{inputId, label, multiple, accept, width}{see \code{shiny::fileInput}}

\item{type}{button type, such as `primary`, `default`, `warning`, `info`,}

\item{class}{additional class to button}

\item{...}{ignored}
}
\description{
Re-write shiny `fileInput`, but minimal
}
