% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{finalize_installation}
\alias{finalize_installation}
\title{Finalize installation}
\usage{
finalize_installation(packages, upgrade = c("always", "ask", "never"))
}
\arguments{
\item{packages}{package name to finalize. 'rave' to only update base demo 
data, or \code{c('threeBrain', 'ravebuiltins')} to upgrade built-in data, 
or leave it blank to upgrade all.}

\item{upgrade}{whether to ask. Default is 'always' to receive default 
settings Other choices are 'ask' or 'never'.}
}
\description{
download demo data
}
