% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_devs.R
\name{find_modules}
\alias{find_modules}
\title{Check all packages to for new RAVE module packages}
\usage{
find_modules(packages, as_module = TRUE)
}
\description{
Check all packages to for new RAVE module packages
}
