% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{fprintf}
\alias{fprintf}
\title{Format Print Strings}
\usage{
fprintf(..., collapse = "\\n", lookup_env = parent.frame())
}
\arguments{
\item{...}{characters}

\item{collapse}{character to collapse characters}

\item{lookup_env}{which environment to look for data?}
}
\description{
Format Print Strings
}
