% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_io.R
\name{freesurfer_mesh}
\alias{freesurfer_mesh}
\title{Function to parse SUMA spec file and generate threejsr free mesh object}
\usage{
freesurfer_mesh(subject, spec_file = NULL, state = "pial",
  center = c(0, 0, 30))
}
\arguments{
\item{subject}{Subject class or character}

\item{spec_file}{default is subject_dir/rave/suma/test.spec}

\item{state}{is pial}

\item{center}{Geom center}
}
\description{
Function to parse SUMA spec file and generate threejsr free mesh object
}
