% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_async.R
\name{get_async_result}
\alias{get_async_result}
\title{Get async process result (deprecated)}
\usage{
get_async_result(name, remove = T)
}
\arguments{
\item{name}{async name}

\item{remove}{remove future object?}
}
\description{
Get async process result (deprecated)
}
