% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_3drender.R
\name{get_color}
\alias{get_color}
\title{Function to get color without errors}
\usage{
get_color(col, alpha = F)
}
\arguments{
\item{col}{color name or number or code}

\item{alpha}{include alpha or not}
}
\description{
Function to get color without errors
}
