% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-parse.R
\name{get_content}
\alias{get_content}
\title{Parse 'RAVE' Module Contents}
\usage{
get_content(content, env, evaluate = TRUE, chunks = FALSE)
}
\arguments{
\item{content}{characters, R code to parse into expressions}

\item{env}{environment to parse the code into, please specify the context}

\item{evaluate}{whether to evaluate parse expression into \code{env}}

\item{chunks}{whether to respect code notations and chunk the code into 
separate parts}
}
\value{
See details.
}
\description{
Parse 'RAVE' Module Contents
}
\details{
If \code{"evaluate=TRUE"}, then the parse code will be evaluated 
within \code{env} and returns a logical value: \code{TRUE} means the content has
something, otherwise returns \code{FALSE}

If \code{"evaluate=FALSE"}, returns the parsed expression and add attributes
about the names of each chunk and whether they are asynchronous
}
