% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file_structure.R
\name{get_dir}
\alias{get_dir}
\title{Get all directories that rave uses}
\usage{
get_dir(subject_code, project_name, block_num, mkdirs = NULL, subject_id,
  relative = F)
}
\arguments{
\item{subject_code}{subject_code}

\item{project_name}{project_name}

\item{block_num}{block_num (optional)}

\item{mkdirs}{mkdirs ensure that some directory exists}

\item{subject_id}{subject_id (project_name/subject_code)}

\item{relative}{relative path or absolute to root data dir}
}
\description{
Get all directories that rave uses
}
