% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{get_fake_updated_message}
\alias{get_fake_updated_message}
\title{internally used for debugging functions}
\usage{
get_fake_updated_message(..., .args = list(), .func = NULL)
}
\arguments{
\item{...}{see with_fake_session}

\item{.args}{same as ...}

\item{.func}{function to pass to with_fake_session}
}
\description{
internally used for debugging functions
}
