% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_overall.R
\name{get_from_package}
\alias{get_from_package}
\title{Get function from external packages}
\usage{
get_from_package(f, pkg, internal = FALSE, ifNotFound = NULL,
  check = TRUE)
}
\arguments{
\item{f}{name of function}

\item{pkg}{package name}

\item{internal}{use "::" or ":::"}

\item{ifNotFound}{if package not found, return this value}

\item{check}{check existence of package?}
}
\description{
Get function from external packages
}
