% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{get_mem_usage}
\alias{get_mem_usage}
\title{Get RAM usage}
\usage{
get_mem_usage(modules, data_envir)
}
\arguments{
\item{modules}{which module(s)}

\item{data_envir}{default uses getDefaultDataRepository}
}
\description{
Get RAM usage
}
