% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-modules.R
\name{get_module}
\alias{get_module}
\title{Function to find modules in packages}
\usage{
get_module(package, module_id, local = FALSE, ...)
}
\arguments{
\item{package}{package name to search for modules}

\item{module_id}{(optional) module ID if the package contains multiple modules}

\item{local}{run module locally?}

\item{...}{ignored for compatibility purpose}
}
\description{
Function to find modules in packages
}
