% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-conf.R
\name{get_path}
\alias{get_path}
\title{Safe Way to Access Module Package Files Using Relative Path}
\usage{
get_path(..., mustWork = FALSE, is_directory = FALSE)
}
\arguments{
\item{...}{relative path to the file}

\item{mustWork}{whether the file must exists}

\item{is_directory}{whether required file is a directory}
}
\value{
If you are developing the package, \code{get_path} returns the 
absolute file path, otherwise it uses \code{\link[base]{system.file}} to 
get the file from package library.
}
\description{
Safe Way to Access Module Package Files Using Relative Path
}
