% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file_structure.R
\name{get_projects}
\alias{get_projects}
\title{Find all Available RAVE Projects}
\usage{
get_projects(refresh = TRUE)
}
\arguments{
\item{refresh}{whether to refresh the project list; default is true}
}
\value{
A vector of RAVE project names in characters
}
\description{
Find all Available RAVE Projects
}
