% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_val}
\alias{get_val}
\title{Get value, if value is invalid, then assign value}
\usage{
get_val(x, key = NULL, ..., .invalids = c('null', 'na'))
}
\arguments{
\item{x}{List or variable}

\item{key}{If not NULL, \code{x[[key]]} will be evaluated}

\item{...}{Default value to be returned if x or x$key is invalid}

\item{.invalids}{See ?is_invalid}
}
\description{
Get value, if value is invalid, then assign value
}
