% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{getDefaultCacheEnvironment}
\alias{getDefaultCacheEnvironment}
\title{Get Cache Environment}
\usage{
getDefaultCacheEnvironment(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{internally used}
}
\description{
Get Cache Environment
}
