% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-cls_module.R
\name{getDefaultDataRepository}
\alias{getDefaultDataRepository}
\title{Get environment where subject data is loaded}
\usage{
getDefaultDataRepository(
  session = getDefaultReactiveDomain(),
  session_id,
  session_based = NULL
)
}
\arguments{
\item{session}{shiny session, default is NULL}

\item{session_id}{internal use}

\item{session_based}{internal use}
}
\description{
Get environment where subject data is loaded
}
