% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_module.R
\name{getDefaultReactiveDomain}
\alias{getDefaultReactiveDomain}
\title{Wrapper for shiny::getDefaultReactiveDomain}
\usage{
getDefaultReactiveDomain()
}
\description{
Wrapper for shiny::getDefaultReactiveDomain
}
