% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-core.R
\name{getExecEnvirFromContext}
\alias{getExecEnvirFromContext}
\title{Get Module Runtime Environment from Current Context}
\usage{
getExecEnvirFromContext()
}
\value{
An \code{\link[rave]{ExecEnvir}} instance
}
\description{
Get Module Runtime Environment from Current Context
}
