% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-core.R
\name{getModuleEnvirFromContext}
\alias{getModuleEnvirFromContext}
\title{Get Module Instance from Current Context}
\usage{
getModuleEnvirFromContext()
}
\value{
An \code{\link[rave]{ModuleEnvir}} instance
}
\description{
Get Module Instance from Current Context
}
