% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{\%?<-\%}
\alias{\%?<-\%}
\title{Assign if not exists, or NULL}
\usage{
lhs \%?<-\% value
}
\description{
Assign if not exists, or NULL
}
\examples{
\dontrun{
# Remove a if exists
if(exists('a', envir = globalenv()))  rm(a, envir = globalenv())

# Assign
a \%?<-\% 1; print(a)

# However, if assigned, nothing happens
a = 1;
a \%?<-\% 2;
print(a)

# in a list
a = list()
a$e \%?<-\% 1; print(a$e)
a$e \%?<-\% 2; print(a$e)
}

}
