% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_map.R
\name{has.class}
\alias{has.class}
\title{Check object has classes}
\usage{
has.class(obj, class, all.match = T, element.wise = F)
}
\arguments{
\item{obj}{object}

\item{class}{vector of classes}

\item{all.match}{all classes needs to be included or only need one match?}

\item{element.wise}{default FALSE}
}
\description{
Check object has classes
}
