% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-brain_model.R
\name{import_electrodes}
\alias{import_electrodes}
\title{Import \code{.csv} files that contain electrode information}
\usage{
import_electrodes(path, subject, use_fs = NA, ...)
}
\arguments{
\item{path}{path to the electrode file to import}

\item{subject}{'RAVE' project-subject combination}

\item{use_fs}{whether to use 'FreeSurfer', default is to auto-detect}

\item{...}{passed to \code{\link[utils]{read.csv}}}
}
\description{
The table to import must contains a column \code{'Electrode'}
that is consistent with the corresponding subject.
}
