% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_devs.R
\name{init_dev_environment}
\alias{init_dev_environment}
\title{Function to create rave package development environment}
\usage{
init_dev_environment(package, reload_subject = FALSE, reactive = FALSE)
}
\description{
Function to create rave package development environment
}
