% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-parse.R
\name{init_module}
\alias{init_module}
\title{Initialize 'RAVE' module for debug purpose}
\usage{
init_module(
  module_id,
  debug = FALSE,
  parse_context = c("rave_running", "rave_running_local")
)
}
\arguments{
\item{module_id}{module ID}

\item{debug}{whether to expose all functions to the global environment}

\item{parse_context}{parsing context, for internal use}
}
\description{
Initialize 'RAVE' module for debug purpose
}
\seealso{
\code{\link[rave]{load_rave_module_package}}
}
