% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{is_invalid}
\alias{is_invalid}
\title{Check if value(s) is invalid}
\usage{
is_invalid(x, any = F, .invalids = c('null', 'na'))
}
\arguments{
\item{x}{Values to check}

\item{any}{If TRUE, then it will check if any element in x is invalid,
otherwise, it will check if all element of x is invalid}

\item{.invalids}{Possible choices: 'null', 'na', 'blank'}
}
\description{
Check if value(s) is invalid
}
\examples{
\dontrun{
is_invalid(NULL)

is_invalid(c(NA, 1))

is_invalid(c(NA, 1), any = T)

is_invalid('', .invalids = 'blank')
}
}
