% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkvalid.R
\name{is_valid}
\alias{is_valid}
\title{Check if object is invalid}
\usage{
is_valid(x, nulls = c(""))
}
\arguments{
\item{x}{object}

\item{nulls}{'na' for is.na, '' for is.blank, 'Inf' for is.infinite, default is.null is also checked}
}
\description{
Check if object is invalid
}
