% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{is_within}
\alias{is_within}
\alias{\%within\%}
\title{Check if value is within a data range}
\usage{
is_within(x, ref, strict = FALSE)

x \%within\% ref
}
\description{
Check if value is within a data range
}
\examples{
\dontrun{
a <- 1:10
a[is_within(a, c(2,5))]
a[is_within(a, c(2,5), strict=T)]
a[is_within(a, 2:5)]

a[a \%within\% 2:5]
}
}
