% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{lapply_expr}
\alias{lapply_expr}
\title{Apply each element under "with" clause}
\usage{
lapply_expr <- function(X, expr, wrapper = NULL, env = environment())
}
\description{
Apply each element under "with" clause
}
\details{
The goal of this function is to get rid of ugly "$" operator.
X should be a list of lists, for each elements, \code{expr} will be
evaluated. You can use the name of the lists directly. If the names are not
provided, use \code{.x} as default variable. See examples.
}
\examples{
\dontrun{
# X is a list of named lists, with name of each elements be "a"
X <- replicate(n = 3, list(a = rnorm(10)))
lapply_expr(X, {mean(a)})

# X is a list of unnamed lists, use ".x" as your vairable names
X <- replicate(n = 3, list(rnorm(10)))
lapply_expr(X, {mean(.x)})

# wrapper needs to be a function. It will be applied to the results
# before returning the values.

# example 1: unlist the result to be a vector
X <- replicate(n = 3, list(rnorm(10)))
lapply_expr(X, {mean(.x)}, wrapper = unlist)

# example 2: wrap up html components using htmltools::tags$ul
lapply_expr(1:10, {
  htmltools::tags$li(sprintf('line: \%d', .x))
}, wrapper = htmltools::tags$ul)
}
}
