% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RaveHistory.R
\name{last_entry}
\alias{last_entry}
\title{Obtain last saved value - disk cache}
\usage{
last_entry(key, default, save = F, group = "customized")
}
\arguments{
\item{key}{key name}

\item{default}{if not found, return default}

\item{save}{replace cache}

\item{group}{character, donnot use "main_app" and "main_app2" since they are reserved}
}
\description{
Obtain last saved value - disk cache
}
