% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io_fst.R
\docType{data}
\name{LazyFST}
\alias{LazyFST}
\title{R6 class to load fst files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LazyFST
}
\description{
R6 class to load fst files
}
\keyword{datasets}
