% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io_h5.R
\docType{data}
\name{LazyH5}
\alias{LazyH5}
\title{File IO: HDF5 file wrapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LazyH5
}
\description{
File IO: HDF5 file wrapper
}
\keyword{datasets}
