% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-fst.R
\name{load_fst_or_h5}
\alias{load_fst_or_h5}
\title{Function try to load "fst" file cache, if not found, read "HDF5" file}
\usage{
load_fst_or_h5(
  fst_path,
  h5_path,
  h5_name,
  fst_need_transpose = F,
  fst_need_drop = F,
  ram = F
)
}
\arguments{
\item{fst_path}{\code{fst} cache path}

\item{h5_path}{alternate \code{hdf5} path}

\item{h5_name}{\code{hdf5} data name}

\item{fst_need_transpose}{does \code{fst} data need transpose?}

\item{fst_need_drop}{drop dimensions}

\item{ram}{read to ram?}
}
\description{
Function try to load "fst" file cache, if not found, read "HDF5" file
}
