% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_local_cache.R
\name{load_local_cache}
\alias{load_local_cache}
\title{Load local cache for fast importing voltage, power, and phase}
\usage{
load_local_cache(project_name, subject_code, epoch, time_range,
  frequency_range = NULL, electrodes, referenced = FALSE,
  data_type = "voltage")
}
\description{
Load local cache for fast importing voltage, power, and phase
}
