% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-csv.R
\name{load_meta}
\alias{load_meta}
\title{Load subject meta data}
\usage{
load_meta(meta_type, project_name, subject_code, subject_id, meta_name)
}
\arguments{
\item{meta_type}{electrodes, epochs, time_points, frequencies, references ...}

\item{project_name}{project name}

\item{subject_code}{subject code}

\item{subject_id}{"project_name/subject_code"}

\item{meta_name}{only used if meta_type is epochs or references}
}
\description{
Load subject meta data
}
