% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modules.R
\name{load_modules}
\alias{load_modules}
\title{Load Modules (Questioning Lifecycle)}
\usage{
load_modules(legacy = FALSE)
}
\arguments{
\item{legacy}{load old RAVE scripts? WARNING: they are not supported anymore.}
}
\description{
Load Modules (Questioning Lifecycle)
}
