% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-conf.R
\name{load_rave_module_package}
\alias{load_rave_module_package}
\title{Function to load RAVE module package with UI tools}
\usage{
load_rave_module_package(
  env,
  parse_context = c("rave_module_debug", "rave_running", "rave_running_local")
)
}
\arguments{
\item{env}{environment to load tools}

\item{parse_context}{parsing context}
}
\description{
called internally by \code{\link[rave]{init_module}} or 
other module packages
}
