% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-load_scripts.R
\name{load_scripts}
\alias{load_scripts}
\title{Load scripts that cannot put into package R folder}
\usage{
load_scripts(..., asis = FALSE)
}
\arguments{
\item{...}{script files that are wrapped by \code{\link[rave]{get_path}}, 
or R quasi-quotations wrapped by \code{\link[rlang]{quo}}.}

\item{asis}{if the scripts to be loaded is a file, whether to copy to a
temporary directory when launching 'RAVE'. Usually we set this to be true
to save loading time. However, if your scripts also source other scripts 
in relative path, we recommend setting \code{asis=FALSE} and also load 
additional scripts using this function.}
}
\value{
None, but will source, or run whatever code provided.
}
\description{
Use in \code{comp.R} to load scripts that cannot be put into 
package \code{"R/"} folder. Usually the scripts contains shiny reactive
values that changes dynamically.
}
\details{
This function raises error when running in default contexts, and 
requires debug mode, or run inside of 'RAVE' instance.
}
\seealso{
\code{\link[rave]{rave_context}}, \code{\link[rave]{get_path}},
\code{\link[rlang]{quo}}
}
