% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_map.R
\name{ls}
\alias{ls}
\title{Internal use to fast-list all elements in an environment}
\usage{
ls(envir = parent.frame(), all.names = FALSE, pattern,
  sorted = FALSE, ...)
}
\arguments{
\item{envir}{environment}

\item{all.names}{include hidden objects}

\item{pattern}{see ls}

\item{sorted}{sort name? default FALSE}

\item{...}{pass to base::ls}
}
\description{
Internal use to fast-list all elements in an environment
}
