% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_map.R
\docType{data}
\name{Map}
\alias{Map}
\title{R6 implementation of hashMap (internal use)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Map
}
\description{
R6 implementation of hashMap (internal use)
}
\keyword{datasets}
