% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_channel_inspection.R
\name{mini_channel_inspection}
\alias{mini_channel_inspection}
\title{Function to activate preprocessing-Gadgets (notch inspection)}
\usage{
mini_channel_inspection(process_name, project_name, subject_code, blocks,
  channels, srate)
}
\description{
Function to activate preprocessing-Gadgets (notch inspection)
}
