% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{module_analysis_names}
\alias{module_analysis_names}
\title{Find module analysis names}
\usage{
module_analysis_names(module_id, project_name)
}
\arguments{
\item{module_id}{module id}

\item{project_name}{project name}
}
\description{
Find module analysis names
}
