% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_group_analysis.R
\name{module_analysis_table}
\alias{module_analysis_table}
\title{Get subject exported analysis table}
\usage{
module_analysis_table(project_name, module_id, analysis_name,
  check_valid = FALSE)
}
\arguments{
\item{project_name}{project name}

\item{module_id}{module id}

\item{analysis_name}{user-specified, use module_analysis_names to get all of'em}

\item{check_valid}{check file existence}
}
\description{
Get subject exported analysis table
}
