% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-conf.R
\name{mount_demo_subject}
\alias{mount_demo_subject}
\alias{mount_demo_subject.rave_module_debug}
\alias{mount_demo_subject.rave_running}
\alias{mount_demo_subject.rave_running_local}
\title{Load Demo Subject According to Package Configuration File}
\usage{
mount_demo_subject(
  subject_code,
  project_name,
  force_reload_subject = FALSE,
  ...,
  download_url
)

\method{mount_demo_subject}{rave_module_debug}(
  subject_code,
  project_name,
  force_reload_subject = FALSE,
  ...,
  download_url
)

\method{mount_demo_subject}{rave_running}(...)

\method{mount_demo_subject}{rave_running_local}(...)
}
\arguments{
\item{subject_code}{optional, subject code}

\item{project_name}{optional, project name}

\item{force_reload_subject}{logical, whether to force reload subject even 
if another subject is loaded}

\item{...}{further passed to \code{\link[rave]{rave_prepare}}}

\item{download_url}{optional, web link to subject archive}
}
\value{
None
}
\description{
Load Demo Subject According to Package Configuration File
}
\details{
When debugging the 'RAVE' modules, it loads demo subject for
debugging from according to settings file \code{"inst/rave.yaml"}. 

This function only function properly in \code{'rave_module_debug'} mode.
This means by default it raises errors. In other mode, for example 
\code{'rave_running'}, it does nothing.
}
