% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_map.R
\docType{data}
\name{MVCAdapter}
\alias{MVCAdapter}
\title{A special type of Map}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MVCAdapter
}
\description{
A special type of Map
}
\keyword{datasets}
