% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-notch.R
\name{notch_channel}
\alias{notch_channel}
\title{Filter line noise out from ECoG channels}
\usage{
notch_channel(s, sample_rate, bands = c(60, 120, 180), width = c(1, 2, 2))
}
\arguments{
\item{s}{signal, time domain}

\item{sample_rate}{signal sample rate}

\item{bands}{bands that will be filtered out}

\item{width}{along with bands, half of the filter width. For example,if bands
is 60Hz and width is 1Hz, then the notch filter lower bound is 60-1=59Hz and
upper bound is 60+1=61Hz.}
}
\description{
Filter line noise out from ECoG channels
}
