% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_notch_filter.R
\name{notch_filter}
\alias{notch_filter}
\title{Notch filter}
\usage{
notch_filter(s, sample_rate, lb, ub, domain = 1)
}
\arguments{
\item{s}{signal in time or frequency domain}

\item{sample_rate}{signal sample rate}

\item{lb}{filter lower bound (Hz)}

\item{ub}{filter upper bound (Hz)}

\item{domain}{1 if the input signal is in the time domain, 0 if it is in the frequency domain}
}
\value{
filtered signal in time domain
}
\description{
Notch filter
}
\details{
This function is alternative R version of notch filter, originally
writen by Mohammad Dastjerdi, Parvizi Lab, Stanford, revised by Brett Foster,
 BCM, Sep 2009, and migrated from Matlab to R by Zhengjia Wang, Rice U, Jan
 2018
}
