% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_observable.R
\docType{data}
\name{Observable}
\alias{Observable}
\title{A template for observables (Abstruct Class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Observable
}
\description{
A template for observables (Abstruct Class)
}
\keyword{datasets}
