% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_observable.R
\docType{data}
\name{Observatory}
\alias{Observatory}
\title{Collection of observables in Observer-Observables pattern
All Observables need to have public method *check()*, and in shiny, those observables will be checked for each \code{interval} time.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Observatory$new(interval = 1000)
}
\arguments{
\item{interval}{Measured in milliseconds, how long shiny should update (check) observables}
}
\description{
Collection of observables in Observer-Observables pattern
All Observables need to have public method *check()*, and in shiny, those observables will be checked for each \code{interval} time.
}
\keyword{datasets}
