% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-options.R
\name{Options}
\alias{Options}
\title{Internal R6 class for rave-options}
\description{
Internal R6 class for rave-options

Internal R6 class for rave-options
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Options$new()}}
\item \href{#method-get_options}{\code{Options$get_options()}}
\item \href{#method-set_options}{\code{Options$set_options()}}
\item \href{#method-reset}{\code{Options$reset()}}
\item \href{#method-load_settings}{\code{Options$load_settings()}}
\item \href{#method-save_settings}{\code{Options$save_settings()}}
\item \href{#method-clone}{\code{Options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$new(conf_path = "~/.rave.yaml", save_default = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_options"></a>}}
\subsection{Method \code{get_options()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$get_options(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_options"></a>}}
\subsection{Method \code{set_options()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$set_options(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_settings"></a>}}
\subsection{Method \code{load_settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$load_settings(conf_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_settings"></a>}}
\subsection{Method \code{save_settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$save_settings(path = "~/.rave.yaml")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
