% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_overall.R
\name{package_installed}
\alias{package_installed}
\title{Check if a package is installed}
\usage{
package_installed(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
logical, if the package is installed, return TRUE, otherwise FALSE
}
\description{
Check if a package is installed
}
