% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parse.R
\name{parse_selections}
\alias{parse_selections}
\title{Convert text to numeric vector}
\usage{
parse_selections(text, sep = "[, ]+", sort = F, unique = T)
}
\arguments{
\item{text}{character like "1-10,13"}

\item{sep}{default is ','}

\item{sort}{sort result}

\item{unique}{remove duplicated value from result?}
}
\description{
Convert text to numeric vector
}
