% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{paste_c}
\alias{paste_c}
\title{Pipe function to paste two characters}
\usage{
paste_c(x, y)
}
\arguments{
\item{x}{character}

\item{y}{character}
}
\value{
paste0(x,y)
}
\description{
Pipe function to paste two characters
}
