% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_units.R
\name{pre_concat}
\alias{pre_concat}
\title{Concatenate final results of wavelet output and perform cumsum
TODO: add process of pdiode data}
\usage{
pre_concat(project_name, subject_code, blocks, chl, srate = 1000,
  target_srate = 100)
}
\description{
Concatenate final results of wavelet output and perform cumsum
TODO: add process of pdiode data
}
