% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_epoch.R
\name{pre_epoch3}
\alias{pre_epoch3}
\title{Preprocess module - epoch}
\usage{
pre_epoch3(module_id = "EPOCH_M", sidebar_width = 2)
}
\arguments{
\item{module_id}{used internally "EPOCH_M"}

\item{sidebar_width}{1 to 12 width for sidebar}
}
\description{
Preprocess module - epoch
}
