% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_raw_signal.R
\name{pre_import_matlab}
\alias{pre_import_matlab}
\title{mport raw voltage signals from matlab file}
\usage{
pre_import_matlab(subject_code, project_name, block_num, chl,
  name = "analogTraces")
}
\arguments{
\item{subject_code}{subject code}

\item{project_name}{project name}

\item{block_num}{which block of signals will be looked into?}

\item{chl}{integer, channel number}

\item{name}{which matlab variable to look at, default is "analogTraces"}
}
\value{
raw signal
}
\description{
mport raw voltage signals from matlab file
}
