% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_load_settings.R
\name{pre_load_settings}
\alias{pre_load_settings}
\title{Deprecated this function was designed for version A,B,C}
\usage{
pre_load_settings(subject_code, setting_file = "variables.txt")
}
\arguments{
\item{subject_code}{subject_code}

\item{setting_file}{setting_file}
}
\description{
Deprecated this function was designed for version A,B,C
}
