% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_signal.R
\name{pwelch}
\alias{pwelch}
\title{Welch Periodogram}
\usage{
pwelch(x, fs, window = 64, noverlap = 8, nfft = 256, col = "black",
  xlim = NULL, ylim = NULL, main = "Welch periodogram",
  plot = TRUE, log = "xy", spec_func = spectrum, cex = 1, ...)
}
\arguments{
\item{x}{signal}

\item{fs}{sample rate}

\item{window}{window length, default 128}

\item{noverlap}{overlap between two adjacent windows, by default is 8}

\item{nfft}{number of DFT points}

\item{col, xlim, ylim, main, cex, ...}{will be passed to plot}

\item{plot}{logical, plot the result or not}

\item{log}{'', 'x', 'y', 'xy', indicates which axis should be log10 value.}

\item{spec_func}{deprecated}
}
\description{
Welch Periodogram
}
\details{
This function is modified from \code{pwelch} function from "oce"
package and is designed for visualizing ECoG voltage data.
}
