% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_python.R
\name{py_console}
\alias{py_console}
\title{Launch python console within R}
\usage{
py_console(compiler_path = "",
  virtualenv = rave_options("py_virtualenv"), py3 = TRUE, ...)
}
\description{
Launch python console within R
}
