% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Subject.R
\name{r_to_py.Subject}
\alias{r_to_py.Subject}
\title{Conver subject to python object}
\usage{
r_to_py.Subject(obj, convert = FALSE)
}
\arguments{
\item{obj}{Subject class}

\item{convert}{pass to reticulate::r_to_py}
}
\description{
Conver subject to python object
}
