% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Tensor.R
\name{r_to_py.Tensor}
\alias{r_to_py.Tensor}
\title{Convert tensor to python objects}
\usage{
r_to_py.Tensor(x, convert = FALSE)
}
\arguments{
\item{x}{Tensor instance}

\item{convert}{TRUE to automatically convert Python objects to their R equivalent. Not implemented}
}
\description{
Convert tensor to python objects
}
