% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{rave-cache}
\alias{rave-cache}
\alias{cache}
\alias{clear_cache}
\alias{remove_cache}
\alias{has_cache}
\title{Cache R Objects with Different levels}
\usage{
cache(
  key,
  val,
  name,
  module_id,
  replace = FALSE,
  across_instance = TRUE,
  across_module = TRUE,
  persist = FALSE
)

clear_cache(levels = 1, module_id)

remove_cache(name, module_id)

has_cache(name, module_id, key)
}
\arguments{
\item{key}{any R object, a named list would be the best.}

\item{val}{value to cache, if key exists, then value will not be evaluated nor saved}

\item{name}{name of the dataset}

\item{module_id}{character, used if object not shared across modules}

\item{replace}{if true, force replace the cached object with current one}

\item{across_instance}{logical, whether can be accessed across shiny session}

\item{across_module}{logical, whether can be accessed across 'RAVE' modules}

\item{persist}{logical, whether persist on the hard-disk}

\item{levels}{levels when clear the cache}
}
\description{
Cache R Objects with Different levels
}
\examples{

print(cache('a', 1, name = 'data')) # returns 1
print(cache('a', 2, name = 'data')) # still returns 1

# clear cache
clear_cache(levels = 3)
print(cache('a', 2, name = 'data')) # Now returns 2

# Not run because a is cached
print(cache('a', 2, name = 'data'))
print(cache('a', {Sys.sleep(10); 1}, name = 'data'))

# Persist on hard-disk
cache('aa', 1:1000000, name = 'junk', replace = TRUE, persist = TRUE)
has_cache('junk')                   # Has data called 'junk'
has_cache('junk', key = 'aaa')      # but no data called 'junk' matching the key
}
