% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\docType{package}
\name{rave-package}
\alias{rave}
\alias{rave-package}
\title{rave: R Analysis and Visualization of 'ECOG/iEEG' Data}
\description{
R Analysis and Visualization of 'ECOG/iEEG' Data RAVE provides cross-platform analysis and visualizations for neuroscience 'ECoG/iEEG' data by "just clicking buttons". This package include: (a) pre-processing of 'ECoG/iEEG' raw data into 'HDF5' format (a cross-platform data format that can be used for R, Matlab, Python, Java and C/C++); (b) 'ECoG/iEEG' epoch and visualizations in real time; (c) Users can write customized analysis as R packages and easily insert into RAVE for interactive visualizations. (d) 3D viewers driven by 'WebGL'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rave.wiki}
  \item \url{https://beauchamplab.github.io/rave/}
  \item Report bugs at \url{https://github.com/beauchamplab/rave/issues}
}

}
\author{
\strong{Maintainer}: Zhengjia Wang \email{dipterix.wang@gmail.com}

Authors:
\itemize{
  \item Zhengjia Wang \email{dipterix.wang@gmail.com}
  \item John Magnotti
  \item Xiang Zhang
  \item Michael S. Beauchamp
}

Other contributors:
\itemize{
  \item Beauchamp lab \email{beauchamplab@gmail.com} [copyright holder]
}

}
\keyword{internal}
