% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-main_app.R
\name{rave-tabs}
\alias{rave-tabs}
\alias{close_tab}
\alias{open_tab}
\title{Open/Close a tab in RAVE main application}
\usage{
close_tab(module_id, tabname)

open_tab(module_id, tabname)
}
\arguments{
\item{module_id}{character, module ID}

\item{tabname}{character, tab box title}
}
\description{
Open/Close a tab in RAVE main application
}
