% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_brain2.R
\name{rave_brain2}
\alias{rave_brain2}
\title{Tools to load and view brain in 3D viewer}
\usage{
rave_brain2(subject, surfaces = "pial", use_141 = TRUE,
  compute_template = FALSE)
}
\arguments{
\item{surfaces}{one or more from `pial`, `white`, `smoothwm`, brain surface types}
}
\description{
Tools to load and view brain in 3D viewer
}
