% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{rave_checks}
\alias{rave_checks}
\title{Check if data is loaded for current module}
\usage{
rave_checks(..., data = NULL)
}
\arguments{
\item{...}{see defails}

\item{data}{same as ..., but can be a vector}
}
\description{
Check if data is loaded for current module
}
\details{
This function checks whether ECoG data is loaded. The format is: DATA+(blankspace)+TYPE. DATA can be
"power" (Wavelet transform amplitude), "phase" (Complex angle), or "volt"/"voltage" (Before wavelet). TYPE can
be "raw" (no reference), "referenced" (referenced by common average reference, white matter reference, or bipolar reference).
For voltage data, there is one more special type "full" which loads voltage data for all electrodes.
}
