% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rave_context_generics}
\alias{rave_context_generics}
\title{Create S3 Generics that Respects 'RAVE' Context}
\usage{
rave_context_generics(fun_name, fun = function() {
})
}
\arguments{
\item{fun_name}{generic function name}

\item{fun}{function that set the arguments of the generic}
}
\value{
A generic function
}
\description{
Create S3 Generics that Respects 'RAVE' Context
}
