% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{rave_failure}
\alias{rave_failure}
\title{RAVE Failure Message}
\usage{
rave_failure(message, level = "ERROR", call = NULL, .stop = TRUE)
}
\arguments{
\item{message}{error message, character}

\item{level}{level of error message; can be chosen from \code{"INFO"},
\code{"WARNING"}, or \code{"ERROR"}}

\item{call}{call expression}

\item{.stop}{stop or just return the condition}
}
\value{
Error condition or stop
}
\description{
RAVE Failure Message
}
