% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{rave_import_rawdata}
\alias{rave_import_rawdata}
\title{Import Raw Signal from Non-standard Formats}
\usage{
rave_import_rawdata(subject_code, project_name, launch_preprocess = TRUE)
}
\arguments{
\item{subject_code}{subject code to search for in the raw folder}

\item{project_name}{project name to create}

\item{launch_preprocess}{whether to launch preprocess app, default is true}
}
\description{
Import Raw Signal from Non-standard Formats
}
