% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lapply.R
\name{rave_lapply}
\alias{rave_lapply}
\title{Async way to lapply with no block}
\usage{
rave_lapply(X, FUN, ..., .get_values = TRUE)
}
\arguments{
\item{X, }{x}

\item{FUN}{see lapply}

\item{...}{see lapply}

\item{.get_values}{return future objects or future values}
}
\description{
Async way to lapply with no block
}
