% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rave_life_cycle_general}
\alias{rave_life_cycle_general}
\title{RAVE development cycle - General}
\description{
RAVE development cycle - General
}
\author{
Zhengjia Wang

There are
\link[rave]{rave_life_cycle_util}
}
