% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rave_life_cycle_signal}
\alias{rave_life_cycle_signal}
\title{RAVE development cycle - Signal Processing}
\section{Part 1 - Transformations}{

\enumerate{
  \item \code{\link{notch_filter}} Matlab-compatible notch filter
  \item \code{\link{wavelet}} Wavelet function compatible with Matlab Fieldtrip package
}
}

\section{Part 2 - Visualizations}{

\enumerate{
  \item \code{\link{diagnose_signal}} Inspect signal(s) with four plots: Time series, 2 x Welch periodogram, Histogram
  \item \code{\link{pwelch}} Welch periodogram for signals
  \item \code{\link{plot_signals}} Plot and compare signals line by line
}
}

\author{
Zhengjia Wang
}
