% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rave_life_cycle_util}
\alias{rave_life_cycle_util}
\title{RAVE development cycle - Utils}
\section{Part 1 - Operators}{

\enumerate{
  \item \code{\link{\%&\%}} Function to concatenate two strings
  \item \code{\link{\%within\%}} Also see \code{\link{is_within}}, check if x is within a range
  \item \code{\link[rave:grapes-help-set-grapes]{\%?<-\%}} Assign value to variable if that variable is \code{NULL} or doesn't exist
}
}

\section{Part 2 - Loops, make life easier}{

\enumerate{
  \item \code{\link{lapply_async}} Apply each functions using future package (async version, parallel)
  \item \code{\link{lapply_expr}} Apply and evaluate each expressions within each applied elements
}
}

\section{Part 3 - Non-standard evaluation, and Evnironment operation}{

\enumerate{
  \item \code{\link{clear_env}} Remove all variables within environment
  \item \code{\link{eval_dirty}} Naive evaluate expressions with side effects, also see \code{\link[rlang]{eval_tidy}}
  \item \code{\link{eval_within}} Evaluate function as if it's run within specified environment

}
}

\section{Part 4 - Validity checks, or others}{

\enumerate{
  \item \code{\link{get_val}} Similar to \%?<-\%, get key from list, returns default values if the key is invalid
  \item \code{\link{is_invalid}} Check validity of data
  \item \code{\link{is.blank}} Check if a string is "" - blank
  \item \code{\link{zero_length}} Check if argument(s) has zero-length
  \item \code{\link{safe_str_c}} Try to make any object to string
  \item \code{\link{try_normalizePath}} Convert relative path to absolute path
}
}

\author{
Zhengjia Wang
}
