% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-module_tools.R
\name{rave_module_tools}
\alias{rave_module_tools}
\title{Tools for module writers}
\usage{
rave_module_tools(
  env = NULL,
  data_env = getDefaultDataRepository(),
  quiet = FALSE
)
}
\arguments{
\item{env}{environment to save tools in}

\item{data_env}{rave data repository returned by rave_prepare, internally used}

\item{quiet}{logical}
}
\description{
Tools for module writers
}
