% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_options.R
\name{rave_options}
\alias{rave_options}
\title{Function to change rave-options}
\usage{
rave_options(..., .save = T, launch_gui = T)
}
\arguments{
\item{...}{Key-Value option pairs}

\item{.save}{save to disk? ignored most of the time}

\item{launch_gui}{launch shiny app?}
}
\description{
Function to change rave-options
}
