% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_notch.R
\name{rave_pre_notch3}
\alias{rave_pre_notch3}
\title{Preprocess Module - Notch filter}
\usage{
rave_pre_notch3(module_id = "NOTCH_M", sidebar_width = 2)
}
\arguments{
\item{module_id}{internally used}

\item{sidebar_width}{sidebar width from 1 to 12}
}
\description{
Preprocess Module - Notch filter
}
