% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_overview.R
\name{rave_pre_overview3}
\alias{rave_pre_overview3}
\title{Preprocess Module - Overview}
\usage{
rave_pre_overview3(module_id = "OVERVIEW_M", sidebar_width = 2)
}
\arguments{
\item{module_id}{internally used}

\item{sidebar_width}{sidebar width from 1 to 12}
}
\description{
Preprocess Module - Overview
}
