% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post_reference.R
\name{rave_pre_postref3}
\alias{rave_pre_postref3}
\title{Preprocess Module - post-reference (deprecated)}
\usage{
rave_pre_postref3(module_id = "POSTREF_M", sidebar_width = 2)
}
\arguments{
\item{module_id}{internally used}

\item{sidebar_width}{sidebar width from 1 to 12}
}
\description{
Preprocess Module - post-reference (deprecated)
}
