% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_reference.R
\name{rave_pre_ref3}
\alias{rave_pre_ref3}
\title{Preprocess Module - Reference (Deprecated)}
\usage{
rave_pre_ref3(module_id = "REF_M", sidebar_width = 2)
}
\arguments{
\item{module_id}{internally used}

\item{sidebar_width}{sidebar width from 1 to 12}
}
\description{
Preprocess Module - Reference (Deprecated)
}
