% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-preprocess.R
\name{rave_preprocess_tools}
\alias{rave_preprocess_tools}
\title{Function to create RAVE preprocess tools}
\usage{
rave_preprocess_tools(env = new.env(), ...)
}
\arguments{
\item{env}{environment to save tools to}

\item{...}{ignored}
}
\description{
Function to create RAVE preprocess tools
}
