% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_options.R
\name{rave_setup}
\alias{rave_setup}
\title{Default setup for rave, mainly create forked sub-processes}
\usage{
rave_setup(func = NULL)
}
\arguments{
\item{func}{NULL by default, function to be added to default (dangerous)}
}
\description{
Default setup for rave, mainly create forked sub-processes
}
