% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_options.R
\name{rave_setup_workers}
\alias{rave_setup_workers}
\title{Default setup for rave, mainly create forked sub-processes}
\usage{
rave_setup_workers(n_workers = NULL, ignore_error = TRUE)
}
\arguments{
\item{n_workers}{NULL by default, maximum number of clusters for parallel}

\item{ignore_error}{whether ignore the errors}
}
\description{
Default setup for rave, mainly create forked sub-processes
}
