% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_brain.R
\docType{data}
\name{RaveBrain}
\alias{RaveBrain}
\title{Class for Brain Mesh/Electrodes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RaveBrain
}
\description{
Class for Brain Mesh/Electrodes
}
\keyword{datasets}
