% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RaveHistory.R
\name{RAVEHistory}
\alias{RAVEHistory}
\title{R6 class for small object caching}
\description{
R6 class for small object caching

R6 class for small object caching
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RAVEHistory$new()}}
\item \href{#method-save}{\code{RAVEHistory$save()}}
\item \href{#method-load}{\code{RAVEHistory$load()}}
\item \href{#method-get_or_save}{\code{RAVEHistory$get_or_save()}}
\item \href{#method-clear}{\code{RAVEHistory$clear()}}
\item \href{#method-clone}{\code{RAVEHistory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEHistory$new(path = "~/", name = ".rave.history.yaml", use_yaml = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEHistory$save(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEHistory$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_or_save"></a>}}
\subsection{Method \code{get_or_save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEHistory$get_or_save(key, val = NULL, save = TRUE, inherits = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEHistory$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEHistory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
