% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_install.R
\name{ravepy}
\alias{ravepy}
\alias{ravepy_remove}
\alias{ravepy_check}
\alias{ravepy_register}
\alias{ravepy_virtualenv_install}
\alias{ravepy_conda_install}
\alias{ravepy_info}
\title{Manage Python Environment from RAVE}
\usage{
ravepy_remove(conda = TRUE)

ravepy_check()

ravepy_register()

ravepy_virtualenv_install(python_path = NULL)

ravepy_conda_install(python_path = NULL)

ravepy_info(quiet = FALSE)
}
\arguments{
\item{conda}{logical, is your environment created using \code{conda}?}

\item{python_path}{character, python path, auo-detected if \code{NULL}. Please specify \code{Python3} path}

\item{quiet}{logical, suppress printing information.}
}
\description{
Manage Python Environment from RAVE
}
