% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_threejsr.R
\name{read.freesurf.asc}
\alias{read.freesurf.asc}
\title{Wrapper for "read.freesurf.asc" (threejsr)}
\usage{
read.freesurf.asc(file)
}
\arguments{
\item{file}{freesurf file path}
}
\description{
Wrapper for "read.freesurf.asc" (threejsr)
}
