% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_threejsr.R
\name{read.freesurf.gii}
\alias{read.freesurf.gii}
\title{Wrapper for "read.freesurf.gii" (threejsr)}
\usage{
read.freesurf.gii(file)
}
\arguments{
\item{file}{freesurf file path}
}
\description{
Wrapper for "read.freesurf.gii" (threejsr)
}
