% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-afnisuma.R
\name{read_mgrid}
\alias{read_mgrid}
\title{Make 'iElvis' mgrid file}
\usage{
read_mgrid(con, raw = F)
}
\arguments{
\item{con}{mgrid file}

\item{raw}{raw file or processed I guess}
}
\description{
Make 'iElvis' mgrid file
}
\author{
J.Magnnotti
}
