% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-ravepipeline.R,
%   R/reexport-ravetools.R, R/rexport-ravecore.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pipeline}
\alias{pipeline_from_path}
\alias{pipeline_list}
\alias{raveio_setopt}
\alias{raveio_getopt}
\alias{load_snippet}
\alias{list_snippets}
\alias{install_snippet}
\alias{logger}
\alias{shift_array}
\alias{baseline_array}
\alias{as_rave_project}
\alias{has_rave_subject}
\alias{as_rave_subject}
\alias{new_rave_subject}
\alias{archive_subject}
\alias{install_subject}
\alias{install_openneuro}
\alias{rave_brain}
\alias{prepare_subject_bare0}
\alias{prepare_subject_bare}
\alias{prepare_subject_voltage_with_epochs}
\alias{prepare_subject_voltage_with_blocks}
\alias{prepare_subject_raw_voltage_with_epochs}
\alias{prepare_subject_raw_voltage_with_blocks}
\alias{prepare_subject_time_frequency_coefficients_with_epochs}
\alias{prepare_subject_time_frequency_coefficients_with_blocks}
\alias{prepare_subject_power_with_epochs}
\alias{prepare_subject_power_with_blocks}
\alias{prepare_subject_phase_with_epochs}
\alias{prepare_subject_phase_with_blocks}
\alias{load_meta2}
\alias{save_meta2}
\alias{collapse_power}
\alias{collapse2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ravecore}{\code{\link[ravecore:archive_subject]{archive_subject()}}, \code{\link[ravecore:as_rave_project]{as_rave_project()}}, \code{\link[ravecore:as_rave_subject]{as_rave_subject()}}, \code{\link[ravecore:collapse2]{collapse2()}}, \code{\link[ravecore:collapse_power]{collapse_power()}}, \code{\link[ravecore:has_rave_subject]{has_rave_subject()}}, \code{\link[ravecore:install_openneuro]{install_openneuro()}}, \code{\link[ravecore:install_subject]{install_subject()}}, \code{\link[ravecore:load_meta2]{load_meta2()}}, \code{\link[ravecore:new_rave_subject]{new_rave_subject()}}, \code{\link[ravecore:prepare_subject_bare]{prepare_subject_bare()}}, \code{\link[ravecore:prepare_subject_bare0]{prepare_subject_bare0()}}, \code{\link[ravecore:prepare_subject_phase_with_blocks]{prepare_subject_phase_with_blocks()}}, \code{\link[ravecore:prepare_subject_phase_with_epochs]{prepare_subject_phase_with_epochs()}}, \code{\link[ravecore:prepare_subject_power_with_blocks]{prepare_subject_power_with_blocks()}}, \code{\link[ravecore:prepare_subject_power_with_epochs]{prepare_subject_power_with_epochs()}}, \code{\link[ravecore:prepare_subject_raw_voltage_with_blocks]{prepare_subject_raw_voltage_with_blocks()}}, \code{\link[ravecore:prepare_subject_raw_voltage_with_epochs]{prepare_subject_raw_voltage_with_epochs()}}, \code{\link[ravecore:prepare_subject_time_frequency_coefficients_with_blocks]{prepare_subject_time_frequency_coefficients_with_blocks()}}, \code{\link[ravecore:prepare_subject_time_frequency_coefficients_with_epochs]{prepare_subject_time_frequency_coefficients_with_epochs()}}, \code{\link[ravecore:prepare_subject_voltage_with_blocks]{prepare_subject_voltage_with_blocks()}}, \code{\link[ravecore:prepare_subject_voltage_with_epochs]{prepare_subject_voltage_with_epochs()}}, \code{\link[ravecore:rave_brain]{rave_brain()}}, \code{\link[ravecore:save_meta2]{save_meta2()}}}

  \item{ravepipeline}{\code{\link[ravepipeline:install_snippet]{install_snippet()}}, \code{\link[ravepipeline:list_snippets]{list_snippets()}}, \code{\link[ravepipeline:load_snippet]{load_snippet()}}, \code{\link[ravepipeline:logger]{logger()}}, \code{\link[ravepipeline:pipeline]{pipeline()}}, \code{\link[ravepipeline:pipeline_from_path]{pipeline_from_path()}}, \code{\link[ravepipeline:pipeline_list]{pipeline_list()}}, \code{\link[ravepipeline:raveio_getopt]{raveio_getopt()}}, \code{\link[ravepipeline:raveio_setopt]{raveio_setopt()}}}

  \item{ravetools}{\code{\link[ravetools:baseline_array]{baseline_array()}}, \code{\link[ravetools:shift_array]{shift_array()}}}
}}

