% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{register_compoundInput}
\alias{register_compoundInput}
\title{Function to register compound inputs to shiny}
\usage{
register_compoundInput()
}
\description{
Function to register compound inputs to shiny
}
