% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-conf.R
\name{reload_module_package}
\alias{reload_module_package}
\title{Reload 'RAVE' module package without restarting 'RStudio'}
\usage{
reload_module_package(expose = FALSE, clear_env = FALSE)
}
\arguments{
\item{expose}{whether to expose development tools to the global environment;
default is no}

\item{clear_env}{whether to clear the global environment before reloading;
default is no}
}
\description{
For debugging module packages. In all other contexts it will
raise error.
}
