% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_threejsr.R
\name{renderThreejs}
\alias{renderThreejs}
\title{Wrapper for "renderThreejs" (threejsr)}
\usage{
renderThreejs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{R expression}

\item{env}{environment to evaluate expr}

\item{quoted}{is expr quoted}
}
\description{
Wrapper for "renderThreejs" (threejsr)
}
