% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{restart_rave}
\alias{restart_rave}
\title{Function to Restart RAVE}
\usage{
restart_rave(reload = T, quiet = FALSE)
}
\arguments{
\item{reload}{reload RAVE after restarting R}

\item{quiet}{logical}
}
\description{
Function to Restart RAVE
}
