% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mem.R
\name{safe_env_size}
\alias{safe_env_size}
\title{Get environment size (deprecated)}
\usage{
safe_env_size(env, sum_up = TRUE)
}
\arguments{
\item{env}{environment}

\item{sum_up}{sum up sizes}
}
\description{
Get environment size (deprecated)
}
