% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_meta.R
\name{safe_read_csv}
\alias{safe_read_csv}
\title{Read csv, with column classes specified}
\usage{
safe_read_csv(file, header = TRUE, sep = ",", colClasses = NA,
  skip = 0, quote = "\\"", ...)
}
\description{
Read csv, with column classes specified
}
