% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{safe_str_c}
\alias{safe_str_c}
\title{Convert to string and never goes wrong}
\usage{
safe_str_c(x, sep = '', collapse = NULL, .error = '')
}
\arguments{
\item{.error}{If x can't be converted to string, return this message}
}
\description{
Convert to string and never goes wrong
}
\examples{
\dontrun{
safe_str_c('Count - ', 3:1)

safe_str_c('Count - ', 3:1, collapse = '..., ')

# aaa doesn't exist
safe_str_c('Count - ', aaa, .error = 'aaa not exists')

# aaa exists
aaa <- 0
safe_str_c('Count - ', aaa, .error = 'aaa not exists')
}
}
