% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_meta.R
\name{safe_write_csv}
\alias{safe_write_csv}
\title{Save data to csv, if file exists, rename old file}
\usage{
safe_write_csv(data, file, ..., quiet = F)
}
\arguments{
\item{data}{data frame}

\item{file}{csv file to save}

\item{...}{pass to write.csv}

\item{quiet}{suppress overwrite message}
}
\description{
Save data to csv, if file exists, rename old file
}
