% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-csv.R
\name{save_meta}
\alias{save_meta}
\title{Function to save meta data to subject}
\usage{
save_meta(data, meta_type, project_name, subject_code)
}
\arguments{
\item{data}{data table}

\item{meta_type}{see load meta}

\item{project_name}{project name}

\item{subject_code}{subject code}
}
\description{
Function to save meta data to subject
}
