% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_widgets.R
\name{section_input}
\alias{section_input}
\title{Tabulize Input Groups}
\usage{
section_input(inputId, label, selected = NULL, ...)
}
\arguments{
\item{inputId}{ID (unique), characters, then in the module, you can use param$this_id to get selected tab name.}

\item{label}{Label name for this component}

\item{selected}{Which tab should be selected when lauch the module. By default NULL means the first tab.}

\item{...}{name=value list, see example}
}
\description{
Input component for different tabs/categories
}
