% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_daemon.R
\name{send_to_daemon}
\alias{send_to_daemon}
\title{Send expressions to execute in daemon process (for side viewers)}
\usage{
send_to_daemon(expr, outputId, type = "threeBrain", save = NULL)
}
\arguments{
\item{expr}{expression to run}

\item{outputId}{shiny outputId}

\item{type}{currently only support 'threeBrain',}

\item{save}{names of objects to be save for daemon process}
}
\description{
Send expressions to execute in daemon process (for side viewers)
}
\details{
This is for module writers only. If you want to show outputs in an
additional window, use this function. RAVE will initialize another process
in the background to communicate with the main process. The daemon process
doesn't share with the original process, hence you need to save whatever
data to be used via `save` argument.
}
