% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-input.R, R/generics-output.R
\name{session-reactives}
\alias{session-reactives}
\alias{getDefaultReactiveInput}
\alias{getDefaultReactiveInput.default}
\alias{getDefaultReactiveInput.rave_module_debug}
\alias{getDefaultReactiveInput.rave_running}
\alias{getDefaultReactiveInput.rave_running_local}
\alias{getDefaultReactiveOutput}
\alias{getDefaultReactiveOutput.default}
\alias{getDefaultReactiveOutput.rave_module_debug}
\alias{getDefaultReactiveOutput.rave_running}
\alias{getDefaultReactiveOutput.rave_running_local}
\title{Get `shiny' "input" and "output" objects under current context}
\usage{
getDefaultReactiveInput(session)

\method{getDefaultReactiveInput}{default}(session)

\method{getDefaultReactiveInput}{rave_module_debug}(session)

\method{getDefaultReactiveInput}{rave_running}(session = shiny::getDefaultReactiveDomain())

\method{getDefaultReactiveInput}{rave_running_local}(session)

getDefaultReactiveOutput(session = shiny::getDefaultReactiveDomain())

\method{getDefaultReactiveOutput}{default}(session = shiny::getDefaultReactiveDomain())

\method{getDefaultReactiveOutput}{rave_module_debug}(session = shiny::getDefaultReactiveDomain())

\method{getDefaultReactiveOutput}{rave_running}(session = shiny::getDefaultReactiveDomain())

\method{getDefaultReactiveOutput}{rave_running_local}(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{shiny session instance}
}
\value{
In shiny context, returns special \code{\link[shiny]{reactiveValues}}
that refers to the inputs and outputs of shiny applications. In non-shiny 
contexts, returns a fake environment related to current fake session, 
for debug purpose.
}
\description{
Get `shiny' "input" and "output" objects under current context
}
