% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-shiny.R
\name{set_rave_theme}
\alias{set_rave_theme}
\title{Set and Return RAVE theme}
\usage{
set_rave_theme(
  theme,
  .set_default = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{theme}{\code{"light"} or \code{"dark"}. See details if missing}

\item{.set_default}{whether to save current theme as default, default is no.}

\item{session}{shiny session}
}
\value{
\code{theme} under current context.
}
\description{
Set and Return RAVE theme
}
\details{
RAVE support two themes: "light" mode and "dark" mode. In "light"
mode, the web application background will be light gray and white. In "dark"
mode, the application background will be gray and foreground will be white.

If \code{theme} is missing and RAVE is running as web application, then 
it is set from current session, otherwise, the default theme is retrieved 
from \code{rave_options('default_theme')}. If option \code{"default_theme"}
is missing, then it defaults to "light".
}
\examples{

# Retrieve current theme
get_val(rave_options('default_theme'), default = 'light')

# Set light mode
set_rave_theme('light')
plot(1:10, main = 'test light mode')

# Set dark mode
set_rave_theme('dark')
plot(1:10, main = 'test dark mode')

}
