% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-shinirize.R
\name{shinirize}
\alias{shinirize}
\title{Convert module to objects used in shiny}
\usage{
shinirize(module, session = getDefaultReactiveDomain(), test.mode = TRUE)
}
\arguments{
\item{module}{\code{\link[rave]{ModuleEnvir}} object}

\item{session}{shiny session, default is current shiny session}

\item{test.mode}{passed by \code{\link[rave]{start_rave}} or
\code{\link[rave]{init_app}}}
}
\description{
Convert module to objects used in shiny
}
