% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-data_selector.R
\name{shiny_data_selector}
\alias{shiny_data_selector}
\title{Shiny modal to be used by init_app for data selection}
\usage{
shiny_data_selector(module_id)
}
\arguments{
\item{module_id}{any character containing 0-9 a-z A-Z}
}
\description{
Shiny modal to be used by init_app for data selection
}
