% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{start_rave}
\alias{start_rave}
\alias{start_rave2}
\title{Start 'RAVE' main application}
\usage{
start_rave2(
  host = "127.0.0.1",
  port = NULL,
  launch.browser = TRUE,
  jupyter = FALSE,
  as_job = FALSE,
  ...
)

start_rave(
  host = "127.0.0.1",
  port = NULL,
  launch.browser = TRUE,
  jupyter = FALSE,
  as_job = FALSE,
  ...
)
}
\arguments{
\item{host}{host IP address; default is \code{"127.0.0.1"}}

\item{port}{integer port number; default is random}

\item{launch.browser}{whether to launch browser; default is true}

\item{jupyter}{whether to launch the 'Jupyter' server; default is false}

\item{as_job}{whether to launch in the background as an 'RStudio' job;
available only in 'RStudio'}

\item{...}{passed to \code{\link[ravedash]{start_session}}}
}
\value{
A 'shiny' application object (invisibly when launched as a job).
}
\description{
Start 'RAVE' main application
}
\examples{

if (interactive()) {
 
start_rave()

}


}
\seealso{
\code{\link[ravedash]{start_session}}
}
