% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Subject.R
\name{Subject}
\alias{Subject}
\title{Subject class}
\description{
Subject class

Subject class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-info}{\code{Subject$info()}}
\item \href{#method-print}{\code{Subject$print()}}
\item \href{#method-finalize}{\code{Subject$finalize()}}
\item \href{#method-new}{\code{Subject$new()}}
\item \href{#method-preprocess_info}{\code{Subject$preprocess_info()}}
\item \href{#method-filter_all_electrodes}{\code{Subject$filter_all_electrodes()}}
\item \href{#method-filter_valid_electrodes}{\code{Subject$filter_valid_electrodes()}}
\item \href{#method-has_bad_time_point}{\code{Subject$has_bad_time_point()}}
\item \href{#method-clone}{\code{Subject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$new(project_name, subject_code, reference = NULL, strict = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess_info"></a>}}
\subsection{Method \code{preprocess_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$preprocess_info(key, default = NULL, customized = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_all_electrodes"></a>}}
\subsection{Method \code{filter_all_electrodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$filter_all_electrodes(electrodes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_valid_electrodes"></a>}}
\subsection{Method \code{filter_valid_electrodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$filter_valid_electrodes(electrodes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_bad_time_point"></a>}}
\subsection{Method \code{has_bad_time_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$has_bad_time_point(block, electrode, start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
