% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_subjectinfo.R
\docType{data}
\name{SubjectInfo2}
\alias{SubjectInfo2}
\title{Definition of subject information}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SubjectInfo2
}
\arguments{
\item{project_name}{Subject project}

\item{subject_code}{Subject code (folder)}

\item{blocks}{All possible blocks}

\item{channels}{All possible channels (including bad, epi channels)}

\item{badchan}{Bad channels (will be ignored)}

\item{epichan}{Epilepsy channels (will be ignored)}

\item{exclchan}{Channels counted as valid but excluded from CAR}
}
\description{
Definition of subject information
}
\keyword{datasets}
