% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_tensor.R
\name{subset.Tensor}
\alias{subset.Tensor}
\title{Subset for Tensor Object}
\usage{
\method{subset}{Tensor}(x, ..., .env = parent.frame())
}
\arguments{
\item{x}{object to be subsetted.}

\item{...}{further arguments to be passed to or from other methods.}

\item{.env}{environtment to evaluate in}
}
\description{
Subset for Tensor Object
}
