% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-afnisuma.R
\name{suma_spec_parse}
\alias{suma_spec_parse}
\title{Parse SUMA spec file}
\usage{
suma_spec_parse(subject, spec_file)
}
\arguments{
\item{subject}{Either characters in format like \code{'Project/Subject'} or 
\code{Subject} object created by \code{Subject$new(...)}}

\item{spec_file}{default decided by \code{rave_options('suma_spec_file')}, 
depending on subjects}
}
\description{
Parse SUMA spec file
}
\examples{

subject = 'Demo/YAB'
# or create subject object
subject = Subject$new('Demo', 'YAB')

\dontrun{
# Please download sample subjects first to run
suma_spec_parse(subject)
}
}
