% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{test_hdspeed}
\alias{test_hdspeed}
\title{Function to test local disk speed}
\usage{
test_hdspeed(file_size = 1e+07, quiet = FALSE)
}
\arguments{
\item{file_size}{in bytes, default is 10 MB}

\item{quiet}{should verbose messages be suppressed?}
}
\description{
Function to test local disk speed
}
