% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_threejsr.R
\name{threejs_scene}
\alias{threejs_scene}
\title{Wrapper for "threejs_scene" (threejsr)}
\usage{
threejs_scene(...)
}
\arguments{
\item{...}{passed to threejsr::threejs_scene}
}
\description{
Wrapper for "threejs_scene" (threejsr)
}
