% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_threejsr.R
\name{threejsOutput}
\alias{threejsOutput}
\title{Wrapper for "threejsOutput" (threejsr)}
\usage{
threejsOutput(outputId, ...)
}
\arguments{
\item{outputId}{Output ID}

\item{...}{passed to threejsr::threejsOutput}
}
\description{
Wrapper for "threejsOutput" (threejsr)
}
