% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-viewdebug.R
\name{to_module}
\alias{to_module}
\title{Parse 'RAVE' Module and Returns Parsed Content in Environments}
\usage{
to_module(
  module_id,
  sidebar_width = 3,
  parse_context = c("rave_running_local", "rave_running")
)
}
\arguments{
\item{module_id}{module ID}

\item{sidebar_width}{input sidebar width}

\item{parse_context}{parse context, default is \code{"rave_running_local"}}
}
\description{
Parse 'RAVE' Module and Returns Parsed Content in Environments
}
