% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{to_ram_size}
\alias{to_ram_size}
\title{Convert bytes to KB, MB, GB,...}
\usage{
to_ram_size(s, kb_to_b = 1000)
}
\arguments{
\item{s}{size}

\item{kb_to_b}{1KB=nB, n=1000 by default}
}
\description{
Convert bytes to KB, MB, GB,...
}
