% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mem.R
\name{trim_env}
\alias{trim_env}
\title{Remove large objects within environment}
\usage{
trim_env(env, large_size = -1, recursive = FALSE,
  remove_biggest_only = TRUE)
}
\arguments{
\item{env}{environment to trim}

\item{large_size}{defines large file size}

\item{recursive}{recursively}

\item{remove_biggest_only}{only remove the biggest object}
}
\description{
Remove large objects within environment
}
