% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{ui_register_function}
\alias{ui_register_function}
\title{Internally register function to shiny (to be deprecated)}
\usage{
ui_register_function(key, update_func = NULL, value_field = "value",
  update_value = F, default_args = list())
}
\arguments{
\item{key}{input function}

\item{update_func}{update function}

\item{value_field}{which is value}

\item{update_value}{I forget what's this}

\item{default_args}{default args when initializing}
}
\description{
Internally register function to shiny (to be deprecated)
}
