% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{updateCompoundInput}
\alias{updateCompoundInput}
\title{Function to update cmpoundInput (broken)}
\usage{
updateCompoundInput(session, inputId, to, ...)
}
\arguments{
\item{session}{shiny session}

\item{inputId}{shiny input ID}

\item{to}{extend count}

\item{...}{other params}
}
\description{
Function to update cmpoundInput (broken)
}
