% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{updatePackageVersion}
\alias{updatePackageVersion}
\title{Internally use to auto update RAVE package version
DO NOT use it, it's copied from websites and I'll probably remove this function in the future}
\usage{
updatePackageVersion(packageLocation = ".")
}
\arguments{
\item{packageLocation}{To be documented}
}
\description{
Internally use to auto update RAVE package version
DO NOT use it, it's copied from websites and I'll probably remove this function in the future
}
