% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-viewdebug.R
\name{view_layout}
\alias{view_layout}
\title{Debug-use only, reload package, mount demo subject, and launch shiny app}
\usage{
view_layout(
  module_id,
  sidebar_width = 3,
  launch.browser = TRUE,
  reload = TRUE,
  ...
)
}
\arguments{
\item{module_id}{module ID to debug}

\item{sidebar_width}{input width, from 1 to 11, default is 3}

\item{launch.browser}{whether to launch browser, default is true, other
options are \code{rstudioapi::viewer} or false.}

\item{reload}{whether to reload package first. default is true,}

\item{...}{passed to \code{\link[rave]{init_app}}}
}
\description{
Debug-use only, reload package, mount demo subject, and launch shiny app
}
