% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_wavelet.R
\name{wavelet}
\alias{wavelet}
\title{Wavelet Transformation With Phase}
\usage{
wavelet(data, freqs, srate, wave_num)
}
\description{
Wavelet Transformation With Phase
}
\details{
Wavelet
Function for decomposing time series data into time-frequency
representation (spectral decomposition) using wavelet transform. Employs
Morlet wavelet method (gaussian taper sine wave) to obtain the analytic
signal for specified frequencies (via convolution).
Inputs:
 data - vector of time series to be decomposed
 freqs - vector of center frequencies for decomposition
 srate - sample rate (in Hz)
 wave_num - desired number of cycles in wavelet (typically 5-10).

Translated from Matlab script written by Brett Foster,
Stanford Memory Lab, Feb. 2015

Last edited: Zhengjia Wang - Mar, 2018
Changes:
  1. Uses fftwtools instead of native functions
  manually calculate convolutions
  Save 50% runtime
  2. Adjusted window length
  3. Return power spec instead of spectrum
  4. Use mvfftw instead of fftw, speed up by another 50%
}
