% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-wavelet.R
\name{wavelet_kernels}
\alias{wavelet_kernels}
\title{Returns wavelets to be used for wavelet function}
\usage{
wavelet_kernels(freqs, srate, wave_num)
}
\arguments{
\item{freqs}{vector of center frequencies for decomposition}

\item{srate}{sample rate (in Hz)}

\item{wave_num}{desired number of cycles in wavelet (typically 3-20 for frequencies 2-200).}
}
\description{
Returns wavelets to be used for wavelet function
}
