% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{with_fake_session}
\alias{with_fake_session}
\title{internally used for debugging functions (reactive)}
\usage{
with_fake_session(func, ...)
}
\arguments{
\item{func}{function}

\item{...}{params for function}
}
\description{
internally used for debugging functions (reactive)
}
