% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{zero_length}
\alias{zero_length}
\title{Calculate if length of input(s) is zero}
\usage{
zero_length(..., any = T, na.rm = F)
}
\arguments{
\item{...}{Element(s) to be evaluated in length}

\item{any}{Any element has zero-length? or all elements need to have zero-length}

\item{na.rm}{Should NA be removed before evaluation?}
}
\description{
Calculate if length of input(s) is zero
}
\examples{
# any = TRUE, any element with zero length will yield "TRUE" result
# In this case, expressions c(1) and a==NULL are not evaluated
zero_length(NULL, c(1), a=={Sys.sleep(10)})

# any = FALSE, if any one has non-zero length, return FALSE
# Notice that in these two cases, "Sys.sleep(10)" is not evaluated
zero_length(NULL, c(1), a=={Sys.sleep(10)}, any = F)

# stop('') yields error, which will be counted as invalid/zero-length
zero_length(stop(''))
}
