% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_customize.R
\name{actionButtonStyled}
\alias{actionButtonStyled}
\title{Re-write shiny actionButton to enable styles}
\usage{
actionButtonStyled(inputId, label, icon = NULL, width = NULL,
  type = "default", btn_type = "button", class = "", ...)
}
\arguments{
\item{inputId, label, icon, width}{see \code{shiny::actionButton}}

\item{type}{default, primary, info... see bootstrap}

\item{btn_type}{html tag attribute "type"}

\item{class}{additional classes}

\item{...}{other methods passed to actionButton}
}
\description{
Re-write shiny actionButton to enable styles
}
