% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file_structure.R
\name{arrange_modules}
\alias{arrange_modules}
\title{Update (optional), and check validity of modules}
\usage{
arrange_modules(refresh = FALSE, reset = FALSE, quiet = FALSE)
}
\arguments{
\item{refresh}{check and updates file}

\item{reset}{same as first_time, check module updates, ignored}

\item{quiet}{no overwrite messages}
}
\description{
Update (optional), and check validity of modules
}
