% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_async.R
\name{async}
\alias{async}
\title{Async function to execute code (depricated)}
\usage{
async(expr, name = NULL, envir = parent.frame())
}
\arguments{
\item{expr}{expression for async exe}

\item{name}{see 'details'}

\item{envir}{environment of code}
}
\description{
Async function to execute code (depricated)
}
\details{
If \code{name} is NULL, expression will be evaluated and nothing will be
 returned. However, if name is specified, you can call function
 \code{get_async_result(name)} later to get the results if once evaluated.
 WARNINGS: if your expression contains large object, the memory will be doubled.
 This is because R doesn't support multi-thread and multisession mode will copy
 those objects if they are mentioned in the expression. Therefore please
 optimize your code carefully.
}
\examples{
\dontrun{
# NOT RUN
# Will not block
library(future)

async({
  # block session for 3 secs
  Sys.sleep(5)
  return('Hello world!')
}, name = 'test') ->
  f

# Still running
resolved(f)

# wait for 5 sec
Sys.sleep(5)

# ready!
resolved(f)
get_async_result('test')
# END
}
}
