% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{cache_raw_voltage}
\alias{cache_raw_voltage}
\title{Import Subject "Matlab" File and Create "HDF5" files}
\usage{
cache_raw_voltage(project_name, subject_code, blocks, electrodes, ...)
}
\arguments{
\item{project_name}{project name}

\item{subject_code}{subject code}

\item{blocks}{blocks to be imported}

\item{electrodes}{to be imported (please check file existence before calling this function)}

\item{...}{ignored}
}
\description{
Import Subject "Matlab" File and Create "HDF5" files
}
