% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check and Install RAVE Dependencies}
\usage{
check_dependencies(
  update_rave = TRUE,
  restart = TRUE,
  nightly = FALSE,
  demo_data = FALSE,
  ...
)
}
\arguments{
\item{update_rave}{logical, whether to update RAVE}

\item{restart}{logical, whether to restart `RStudio` after installation}

\item{nightly}{logical, whether to install develop version}

\item{demo_data}{logical, whether to check demo data}

\item{...}{for compatibility purpose, ignored}
}
\description{
Check and Install RAVE Dependencies
}
