% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_syscheck.R
\name{check_updates}
\alias{check_updates}
\title{Check dependencies and update them at start up (Highly recommended)}
\usage{
check_updates(file)
}
\arguments{
\item{file}{file to check package update}
}
\description{
Check dependencies and update them at start up (Highly recommended)
}
