% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{create_local_cache}
\alias{create_local_cache}
\title{Create local cache to speed up loading speed}
\usage{
create_local_cache(project_name, subject_code, epoch, time_range)
}
\arguments{
\item{project_name}{project name}

\item{subject_code}{subject code}

\item{epoch}{epoch name}

\item{time_range}{time range to cache}
}
\description{
Create local cache to speed up loading speed
}
