% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{customizedUI}
\alias{customizedUI}
\title{Customized Shiny Elements}
\usage{
customizedUI(inputId, width = 12L, ...)
}
\arguments{
\item{inputId}{character, input id}

\item{width}{integer from 1-12}

\item{...}{passed to \code{\link[shiny]{uiOutput}}}
}
\description{
Customized Shiny Elements
}
