% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-decimate.R
\name{decimate_fir}
\alias{decimate_fir}
\title{Decimate or downsample a signal using FIR filter}
\usage{
decimate_fir(x, q, n = 30)
}
\arguments{
\item{x}{signal to be decimated}

\item{q}{integer factor to downsample by}

\item{n}{filter order used in the downsampling, default 30}
}
\value{
Downsampled signal
}
\description{
Decimate or downsample a signal using FIR filter
}
