% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-decimate.R
\name{decimate_fir}
\alias{decimate_fir}
\title{Decimate or Down-sample a Signal using \code{FIR} Filters}
\usage{
decimate_fir(x, q, n = 30)
}
\arguments{
\item{x}{signal to be decimated}

\item{q}{integer factor to decimated by}

\item{n}{filter order used for down-sample procedure, default is 30}
}
\value{
Down-sampled signal
}
\description{
Decimate or Down-sample a Signal using \code{FIR} Filters
}
