% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-plots.R
\name{diagnose_signal}
\alias{diagnose_signal}
\title{Plot and Inspect Signals in Trace, Periodogram, and Histogram}
\usage{
diagnose_signal(
  s1,
  s2 = NULL,
  sc = NULL,
  srate,
  name = "",
  try_compress = TRUE,
  max_freq = 300,
  window = ceiling(srate * 2),
  noverlap = window/2,
  std = 3,
  cex = 1.5,
  lwd = 0.5,
  flim = NULL,
  nclass = 100,
  main = "Channel Inspection",
  col = c("black", "red"),
  which = NULL,
  start_time = 0,
  boundary = NULL,
  mar = c(5.2, 5.1, 4.1, 2.1),
  ...
)
}
\arguments{
\item{s1}{Signal for inspection}

\item{s2}{Signal to compare, default NULL}

\item{sc}{compressed signal to speedup the trace plot, if not provided, then 
either the original \code{s1} is used, or a compressed version will be used. 
See parameter \code{try_compress}.}

\item{srate}{Sample rate of s1, note that \code{s2} and \code{s1} must have 
the same sample rate}

\item{name}{Analysis name, for e.g. "CAR", "Notch", etc.}

\item{try_compress}{If length of \code{s1} is too large, it might take long 
to draw trace plot, my solution is to down-sample s1 first (like what Matlab 
does), and then plot the compressed signal. Some information will be lost 
during this process, however, the trade-off is the speed. 
\code{try_compress=FALSE} indicates that you don't want to compress signals 
under any situation (this might be slow).}

\item{max_freq}{Max frequency to plot, should be no larger than half of 
the sampling rate.}

\item{window}{Window length to draw the Periodogram}

\item{noverlap}{Number of data points that each adjacent windows overlap}

\item{std}{Error bar (red line) be drawn at standard deviations, by default 
is 3, meaning the error bars represent 3 standard deviations.}

\item{cex, lwd, mar, ...}{passed to \code{\link[graphics]{plot.default}}}

\item{flim}{\code{log10} of frequency range to plot}

\item{nclass}{Number of classes for histogram}

\item{main}{Plot title}

\item{col}{Color for two signals, length of 2.}

\item{which}{Which sub-plot to plot}

\item{start_time}{When does signal starts}

\item{boundary}{Boundary for signal plot, default is 1 standard deviation}
}
\description{
Plot and Inspect Signals in Trace, Periodogram, and Histogram
}
\examples{
library(stats)
time <- seq(0, 100, by = 1/200)
s2 <- sin(2 * pi * 60 * time) + rnorm(length(time))
diagnose_signal(s2, srate = 200)

# Apply notch filter
s1 = notch_filter(s2, 200, 58,62)
diagnose_signal(s1, s2, srate = 200)


}
