% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ECoGRepository.R
\name{ECoGRepository}
\alias{ECoGRepository}
\title{R6 class for ECoG data Repository}
\description{
R6 class for ECoG data Repository

R6 class for ECoG data Repository
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-info}{\code{ECoGRepository$info()}}
\item \href{#method-print}{\code{ECoGRepository$print()}}
\item \href{#method-new}{\code{ECoGRepository$new()}}
\item \href{#method-get_electrode}{\code{ECoGRepository$get_electrode()}}
\item \href{#method-load_electrodes}{\code{ECoGRepository$load_electrodes()}}
\item \href{#method-epoch}{\code{ECoGRepository$epoch()}}
\item \href{#method-load_reference}{\code{ECoGRepository$load_reference()}}
\item \href{#method-baseline}{\code{ECoGRepository$baseline()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$info(print = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$new(subject, reference = "default", autoload = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_electrode"></a>}}
\subsection{Method \code{get_electrode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$get_electrode(electrode, name = "raw")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_electrodes"></a>}}
\subsection{Method \code{load_electrodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$load_electrodes(electrodes, reference = "default")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epoch"></a>}}
\subsection{Method \code{epoch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$epoch(
  epoch_name,
  pre,
  post,
  electrodes = NULL,
  frequency_range = NULL,
  data_type = "power",
  referenced = T,
  func = NULL,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_reference"></a>}}
\subsection{Method \code{load_reference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$load_reference(ref_name, electrodes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baseline"></a>}}
\subsection{Method \code{baseline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGRepository$baseline(from, to, electrodes = NULL, print.time = FALSE)}\if{html}{\out{</div>}}
}

}
}
