% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ECoGTensor.R
\name{ECoGTensor}
\alias{ECoGTensor}
\title{ECoG Tensor class inherit from \code{\link[rave]{Tensor}}}
\description{
Four-mode tensor (array) especially designed for 
\code{iEEG/ECoG} data. The Dimension names are: \code{Trial}, 
\code{Frequency}, \code{Time}, and \code{Electrode}.
}
\author{
Zhengjia Wang
}
\section{Super class}{
\code{\link[rave:Tensor]{rave::Tensor}} -> \code{ECoGTensor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-flatten}{\code{ECoGTensor$flatten()}}
\item \href{#method-new}{\code{ECoGTensor$new()}}
\item \href{#method-clone}{\code{ECoGTensor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id=".use_multi_files">}\href{../../rave/html/Tensor.html#method-.use_multi_files}{\code{rave::Tensor$.use_multi_files()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="collapse">}\href{../../rave/html/Tensor.html#method-collapse}{\code{rave::Tensor$collapse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="finalize">}\href{../../rave/html/Tensor.html#method-finalize}{\code{rave::Tensor$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="get_data">}\href{../../rave/html/Tensor.html#method-get_data}{\code{rave::Tensor$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="operate">}\href{../../rave/html/Tensor.html#method-operate}{\code{rave::Tensor$operate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="print">}\href{../../rave/html/Tensor.html#method-print}{\code{rave::Tensor$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="remove_data">}\href{../../rave/html/Tensor.html#method-remove_data}{\code{rave::Tensor$remove_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="set_data">}\href{../../rave/html/Tensor.html#method-set_data}{\code{rave::Tensor$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="subset">}\href{../../rave/html/Tensor.html#method-subset}{\code{rave::Tensor$subset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="to_swap">}\href{../../rave/html/Tensor.html#method-to_swap}{\code{rave::Tensor$to_swap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rave" data-topic="Tensor" data-id="to_swap_now">}\href{../../rave/html/Tensor.html#method-to_swap_now}{\code{rave::Tensor$to_swap_now()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-flatten"></a>}}
\if{latex}{\out{\hypertarget{method-flatten}{}}}
\subsection{Method \code{flatten()}}{
converts tensor (array) to a table (data frame)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGTensor$flatten(include_index = TRUE, value_name = "value")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_index}}{logical, whether to include dimension names}

\item{\code{value_name}}{character, column name of the value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame with the dimension names as index columns and 
\code{value_name} as value column
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGTensor$new(
  data,
  dim,
  dimnames,
  varnames,
  hybrid = FALSE,
  swap_file = tempfile(),
  temporary = TRUE,
  multi_files = FALSE,
  use_index = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{array or vector}

\item{\code{dim}}{dimension of data, mush match with \code{data}}

\item{\code{dimnames}}{list of dimension names, equal length as \code{dim}}

\item{\code{varnames}}{names of \code{dimnames}, recommended names are: 
\code{Trial}, \code{Frequency}, \code{Time}, and \code{Electrode}}

\item{\code{hybrid}}{whether to enable hybrid mode to reduce RAM usage}

\item{\code{swap_file}}{if hybrid mode, where to store the data}

\item{\code{temporary}}{whether to clean up the space when exiting R session}

\item{\code{multi_files}}{logical, whether to use multiple files instead of 
one giant file to store data}

\item{\code{use_index}}{logical, when \code{multi_files} is true, whether use 
index dimension as partition number}

\item{\code{...}}{further passed to \code{\link[rave]{Tensor}} constructor}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an \code{ECoGTensor} instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ECoGTensor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
