% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ECoGTensor.R
\docType{data}
\name{ECoGTensor}
\alias{ECoGTensor}
\title{ECoG Tensor class inherit from Tensor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ECoGTensor
}
\description{
ECoG Tensor class inherit from Tensor
}
\keyword{datasets}
