% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Electrode.R
\name{Electrode}
\alias{Electrode}
\title{R6 class of electrode}
\description{
R6 class of electrode

R6 class of electrode
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-info}{\code{Electrode$info()}}
\item \href{#method-print}{\code{Electrode$print()}}
\item \href{#method-switch_reference}{\code{Electrode$switch_reference()}}
\item \href{#method-referenced}{\code{Electrode$referenced()}}
\item \href{#method-clean}{\code{Electrode$clean()}}
\item \href{#method-new}{\code{Electrode$new()}}
\item \href{#method-epoch}{\code{Electrode$epoch()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-switch_reference"></a>}}
\subsection{Method \code{switch_reference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$switch_reference(new_reference)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-referenced"></a>}}
\subsection{Method \code{referenced()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$referenced(type = "power", ram = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\subsection{Method \code{clean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$clean(types = c("power", "phase", "volt"), force = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$new(
  subject,
  electrode,
  reference_by = "noref",
  preload = NULL,
  is_reference = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epoch"></a>}}
\subsection{Method \code{epoch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$epoch(
  epoch_name,
  pre,
  post,
  types = c("volt", "power", "phase"),
  raw = F,
  hybrid = TRUE
)}\if{html}{\out{</div>}}
}

}
}
