% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Electrode.R
\name{Electrode}
\alias{Electrode}
\title{R6 Class for Electrode}
\description{
Stores single electrode or reference signals
}
\examples{
\dontrun{
# Electrode with no reference
e1 <- Electrode$new('demo/YAB', electrode = 14, reference_by = 'noref')
e1$reference
#> Subject: demo/YAB
#> Electrode: noref (Reference)

# Add Common Average Reference in rave/data/reference/ref_13-63,65-84.h5
e2 <- Electrode$new('demo/YAB', electrode = 14, 
                    reference_by = 'ref_13-63,65-84')

# Electrode with bipolar reference by another electrode
e3 <- Electrode$new('demo/YAB', electrode = 14, reference_by = 'ref_15')

# Alternative way
reference <- Electrode$new('demo/YAB', electrode = 15, is_reference = TRUE)
e4 <- Electrode$new('demo/YAB', electrode = 14, reference_by = reference)

# e3, e4 are the same in scientific meaning. To test it, epoch them
power3 <- e3$epoch('YABaOutlier', 1, 2, 'power', 
                   raw = FALSE, hybrid = FALSE)$power
power4 <- e4$epoch('YABaOutlier', 1, 2, 'power',
                   raw = FALSE, hybrid = TRUE)$power

# Compare e3 and e4, result difference should be 0
range(power3$get_data() - power4$get_data())
#> 0

# With or without hybrid, the size will be different
# No hybrid, totally in memory
lobstr::obj_size(power3)
#> 12 MB
# Hybrid, data is swapped to hard-drive
lobstr::obj_size(power4)
#> 908 kB
}
}
\author{
Zhengjia Wang
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{electrode}}{electrode number in integer}

\item{\code{raw_power}}{stores pre-epoch power spectrum with no reference}

\item{\code{raw_phase}}{stores pre-epoch phase with no reference}

\item{\code{raw_volt}}{stores pre-epoch analog traces with no reference}

\item{\code{phase}}{stores pre-epoch phase after reference}

\item{\code{power}}{stores pre-epoch power spectrum after reference}

\item{\code{volt}}{stores pre-epoch analog traces after reference}

\item{\code{preload}}{which of the three data are pre-loaded}

\item{\code{reference}}{character or \code{Electrode} instance indicating the 
reference for current electrode}

\item{\code{has_power, has_phase, has_volt}}{whether power, phase, or voltage
data exist in file system before and after reference}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{has_power, has_phase, has_volt}}{whether power, phase, or voltage
data exist in file system before and after reference}

\item{\code{blocks}}{character vector of block names (read-only)}

\item{\code{subject_id}}{character of subject ID (read-only)}

\item{\code{reference_electrode}}{whether this is a reference (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-info}{\code{Electrode$info()}}
\item \href{#method-print}{\code{Electrode$print()}}
\item \href{#method-switch_reference}{\code{Electrode$switch_reference()}}
\item \href{#method-referenced}{\code{Electrode$referenced()}}
\item \href{#method-clean}{\code{Electrode$clean()}}
\item \href{#method-new}{\code{Electrode$new()}}
\item \href{#method-epoch}{\code{Electrode$epoch()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
print electrode information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
overrides default print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-switch_reference"></a>}}
\if{latex}{\out{\hypertarget{method-switch_reference}{}}}
\subsection{Method \code{switch_reference()}}{
switch reference (experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$switch_reference(new_reference)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_reference}}{An electrode instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-referenced"></a>}}
\if{latex}{\out{\hypertarget{method-referenced}{}}}
\subsection{Method \code{referenced()}}{
get referenced data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$referenced(type = "power", ram = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{which data to reference, default is power}

\item{\code{ram}}{whether to load data to memory}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{ram} is true, then returns a list of matrices. The 
length of the list equals the number of blocks, and each matrix is 
frequency by time points. If \code{ram} is false, then returns an 
environment with each element a \code{\link[raveio]{LazyH5}} or 
\code{\link[raveio]{LazyFST}} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
remove data from memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$clean(types = c("power", "phase", "volt"), force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{types}}{data types to clean}

\item{\code{force}}{whether to remove pre-loaded data types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$new(
  subject,
  electrode,
  reference_by = "noref",
  preload = NULL,
  is_reference = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject}}{\code{\link[rave]{Subject}} instance or characters like 
\code{"proj/sub"}}

\item{\code{electrode}}{number, integer}

\item{\code{reference_by}}{reference signals, choices are character, or 
\code{\link[rave]{Electrode}} instance; default is \code{"noref"}, 
meaning no reference to the electrode}

\item{\code{preload}}{data to load along with constructor}

\item{\code{is_reference}}{is current instance a reference?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{\link[rave]{Electrode}} instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epoch"></a>}}
\if{latex}{\out{\hypertarget{method-epoch}{}}}
\subsection{Method \code{epoch()}}{
epoch electrode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electrode$epoch(
  epoch_name,
  pre,
  post,
  types = c("volt", "power", "phase"),
  raw = FALSE,
  hybrid = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epoch_name}}{epoch name, for example, \code{epoch_name="default"}
refers to epoch file \code{"epoch_default.csv"} in subject meta folder}

\item{\code{pre}}{seconds before trial onset to load}

\item{\code{post}}{seconds after trial onset to load}

\item{\code{types}}{characters, data types to load; choices are \code{"volt"},
\code{"power"}, and \code{"phase"}}

\item{\code{raw}}{whether epoch pre-referenced data?}

\item{\code{hybrid}}{whether to fast-cache the data on hard-drive? See also 
\code{\link[raveio]{Tensor}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of data after epoch
}
}
}
