% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_module_env.R
\docType{data}
\name{ExecEnvir}
\alias{ExecEnvir}
\title{Module runtime environment (session based)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExecEnvir
}
\description{
Module runtime environment (session based)
}
\keyword{datasets}
