% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-plots.R
\name{export_diagnose_voltage}
\alias{export_diagnose_voltage}
\title{Export voltage (analog trace) diagnostic plots for each electrode}
\usage{
export_diagnose_voltage(
  subject,
  electrodes,
  blocks,
  save_dir = "./export",
  width = 12,
  height = 7,
  useDingbats = F,
  onefile = T,
  winlen,
  freq_lim,
  nclass = 50,
  fore_col = "black",
  back_col = "grey80",
  ...
)
}
\arguments{
\item{subject}{character with the following format: 
\code{"project_name/subject_code"}}

\item{electrodes}{a integer vector. For example: \code{c(1,3,4,10:15)}}

\item{blocks}{the blocks to include. Default is all blocks of this subject}

\item{save_dir}{the directory you want to save the files}

\item{width, height, useDingbats}{passed to \code{\link[grDevices]{pdf}}}

\item{onefile}{collect images within one file?}

\item{winlen}{window length, default is twice of the subject sampling rate}

\item{freq_lim}{default is half of voltage sampling rate}

\item{nclass}{number of classes in histogram plot, default is 50}

\item{fore_col}{Periodogram color for Notch filtered signals}

\item{back_col}{Periodogram color for raw signals}

\item{...}{All other parameters passed to \code{\link[grDevices]{pdf}}}
}
\description{
You must import subject through \code{rave_preprocess()} and 
then run this function
}
\seealso{
\code{\link[grDevices]{pdf}}, \code{\link{diagnose_signal}}, 
\code{\link[rave]{pwelch}}
}
