% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{finalize_installation}
\alias{finalize_installation}
\title{Finalize installation}
\usage{
finalize_installation(
  packages,
  upgrade = c("ask", "config-only", "always", "never", "data-only"),
  async = FALSE
)
}
\arguments{
\item{packages}{package name to finalize. 'rave' to only update base demo 
data, or \code{c('threeBrain', 'ravebuiltins')} to upgrade built-in data, 
or leave it blank to upgrade all.}

\item{upgrade}{whether to ask. Default is \code{'always'} to receive default 
settings. Other choices are \code{'ask'}, \code{'never'}, 
\code{'config-only'}, and \code{'data-only'}}

\item{async}{whether to run scripts in parallel; default is true.}
}
\description{
download demo data
}
