% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file_structure.R
\name{get_dir}
\alias{get_dir}
\title{Get Directories in `RAVE`}
\usage{
get_dir(
  subject_code,
  project_name,
  block_num,
  mkdirs = NULL,
  subject_id,
  relative = FALSE
)
}
\arguments{
\item{subject_code}{subject code; can be ignored when \code{subject_id} is provided}

\item{project_name}{project name; can be ignored when \code{subject_id} is provided}

\item{block_num}{block name (optional)}

\item{mkdirs}{internally used}

\item{subject_id}{subject ID; can be omitted if \code{subject_code} and 
\code{project_name} are provided}

\item{relative}{whether to return relative path or absolute to root directory}
}
\description{
Get Directories in `RAVE`
}
