% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{get_mem_usage}
\alias{get_mem_usage}
\title{Get RAM usage}
\usage{
get_mem_usage(
  modules = list(),
  data_envir = getDefaultDataRepository(),
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{modules}{which module(s)}

\item{data_envir}{default uses \code{\link[rave]{getDefaultDataRepository}}}

\item{session}{shiny session instance}
}
\description{
Get RAM usage
}
