% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file_structure.R
\name{get_projects}
\alias{get_projects}
\title{Find all Available RAVE Projects}
\usage{
get_projects()
}
\value{
A vector of RAVE project names in characters
}
\description{
Find all Available RAVE Projects
}
