% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{get_val}
\alias{get_val}
\title{Get Value or Default}
\usage{
get_val(x, key = NULL, ..., .invalids = c("null", "na"))
}
\arguments{
\item{x}{a list, or environment, or just any R object}

\item{key}{the name to obtain from \code{x}. Default is \code{NULL}}

\item{...}{if the value is invalid, the default value to return}

\item{.invalids}{what counts as invalid? Default is \code{NULL} and 
\code{NA}, represented by \code{"null"} and \code{"na"}}
}
\description{
Get Value or Default
}
