% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-debug_app.R
\name{init_app}
\alias{init_app}
\title{Initialize main application for debugging purpose}
\usage{
init_app(
  modules = NULL,
  active_module = NULL,
  launch.browser = TRUE,
  theme = "red",
  disable_sidebar = FALSE,
  simplify_header = FALSE,
  ...,
  data_repo = getDefaultDataRepository()
)
}
\arguments{
\item{modules}{which modules to show. See \code{\link[rave]{load_modules}}}

\item{active_module}{which module to focus at start up (use module ID)}

\item{launch.browser}{launch browsers, default is on}

\item{theme}{color theme for the website}

\item{disable_sidebar}{hide sidebar at startup?}

\item{simplify_header}{hide header at startup?}

\item{...}{other parameters like \code{test.mode} for module debugging}

\item{data_repo}{internally used}
}
\description{
Initialize main application for debugging purpose
}
