% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LazyFST.R
\name{LazyFST}
\alias{LazyFST}
\title{R6 class to load fst files}
\description{
R6 class to load fst files

R6 class to load fst files
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-open}{\code{LazyFST$open()}}
\item \href{#method-close}{\code{LazyFST$close()}}
\item \href{#method-save}{\code{LazyFST$save()}}
\item \href{#method-new}{\code{LazyFST$new()}}
\item \href{#method-get_dims}{\code{LazyFST$get_dims()}}
\item \href{#method-subset}{\code{LazyFST$subset()}}
\item \href{#method-clone}{\code{LazyFST$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$open(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$close(..., .remove_file = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$save(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$new(file_path, transpose = F, dims = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dims"></a>}}
\subsection{Method \code{get_dims()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$get_dims(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$subset(i = NULL, j = NULL, ..., drop = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyFST$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
