% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LazyH5.R
\name{LazyH5}
\alias{LazyH5}
\title{File IO: HDF5 file wrapper}
\description{
File IO: HDF5 file wrapper

File IO: HDF5 file wrapper
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-finalize}{\code{LazyH5$finalize()}}
\item \href{#method-print}{\code{LazyH5$print()}}
\item \href{#method-new}{\code{LazyH5$new()}}
\item \href{#method-save}{\code{LazyH5$save()}}
\item \href{#method-open}{\code{LazyH5$open()}}
\item \href{#method-close}{\code{LazyH5$close()}}
\item \href{#method-subset}{\code{LazyH5$subset()}}
\item \href{#method-get_dims}{\code{LazyH5$get_dims()}}
\item \href{#method-clone}{\code{LazyH5$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$new(file_path, data_name, read_only = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$save(
  x,
  chunk = "auto",
  level = 7,
  replace = TRUE,
  new_file = FALSE,
  force = TRUE,
  ctype = NULL,
  size = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$open(new_dataset = F, robj, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$close(all = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$subset(..., drop = FALSE, stream = F, envir = parent.frame())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dims"></a>}}
\subsection{Method \code{get_dims()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$get_dims(stay_open = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LazyH5$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
