% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-hdf5.R
\name{load_h5}
\alias{load_h5}
\title{Lazy Load "HDF5" File via \code{\link[hdf5r]{hdf5r-package}}}
\usage{
load_h5(file, name, read_only = TRUE, ram = FALSE)
}
\arguments{
\item{file}{"HDF5" file}

\item{name}{\code{group/data_name} path to dataset}

\item{read_only}{default is TRUE, read dataset only}

\item{ram}{load to RAM immediately}
}
\description{
Wrapper for class \code{\link[rave]{LazyH5}}, which load data with 
"lazy" mode - only read part of dataset when needed.
}
\seealso{
\code{\link[rave]{save_h5}}
}
