% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io_h5.R
\name{load_h5}
\alias{load_h5}
\title{Lazy load HDF5 file via rhdf5 package}
\usage{
load_h5(file, name, read_only = T)
}
\description{
Lazy load HDF5 file via rhdf5 package
}
\details{
load_h5 is a wrapper for class LazyH5, which load data with "lazy" mode -
only read part of dataset when needed.
}
\examples{
\dontrun{
f <- system.file('data/data_dir/Subject_RAVE_Demo/rave/cache/6.h5', package = 'rave')
name <- '/wavelet/power/008'
dat <- load_h5(f, name)
dim(dat)
dat[,1:3]
}
}
\seealso{
\code{\link{save_h5}}
}
