% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logger.R
\name{logger}
\alias{logger}
\title{Color console for RAVE}
\usage{
logger(..., level = "DEBUG")
}
\arguments{
\item{...}{things to print}

\item{level}{logger level: DEBUG, INFO, WARNING, ERROR, FATAL}
}
\description{
Color console for RAVE
}
