% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_group_analysis.R
\name{module_analysis_names}
\alias{module_analysis_names}
\title{Find module analysis names}
\usage{
module_analysis_names(module_id, project_name)
}
\arguments{
\item{module}{id}

\item{project}{name}
}
\description{
Find module analysis names
}
