% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_module.R
\docType{data}
\name{ModuleEnvir}
\alias{ModuleEnvir}
\title{Module class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ModuleEnvir
}
\description{
Module class
}
\keyword{datasets}
