% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModuleEnvir.R
\name{ModuleEnvir}
\alias{ModuleEnvir}
\title{Module class}
\description{
Module class

Module class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-info}{\code{ModuleEnvir$info()}}
\item \href{#method-print}{\code{ModuleEnvir$print()}}
\item \href{#method-new}{\code{ModuleEnvir$new()}}
\item \href{#method-get_or_new_exec_env}{\code{ModuleEnvir$get_or_new_exec_env()}}
\item \href{#method-load_script}{\code{ModuleEnvir$load_script()}}
\item \href{#method-cache}{\code{ModuleEnvir$cache()}}
\item \href{#method-render_ui}{\code{ModuleEnvir$render_ui()}}
\item \href{#method-clean}{\code{ModuleEnvir$clean()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$new(
  module_id,
  label_name,
  script_path,
  author = NULL,
  version = "0",
  packages = NULL,
  .script_content = NULL,
  rmd_path = NULL,
  parent_env = globalenv()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_or_new_exec_env"></a>}}
\subsection{Method \code{get_or_new_exec_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$get_or_new_exec_env(
  session = getDefaultReactiveDomain(),
  ...,
  new = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_script"></a>}}
\subsection{Method \code{load_script()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$load_script(session = getDefaultReactiveDomain())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cache"></a>}}
\subsection{Method \code{cache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$cache(key, val, session, replace = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_ui"></a>}}
\subsection{Method \code{render_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$render_ui(session = getDefaultReactiveDomain())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\subsection{Method \code{clean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleEnvir$clean(session = getDefaultReactiveDomain(), session_id)}\if{html}{\out{</div>}}
}

}
}
