% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_options.R
\docType{data}
\name{Options}
\alias{Options}
\title{Internal R6 class for rave-options}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Options
}
\description{
Internal R6 class for rave-options
}
\keyword{datasets}
