% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-plots.R
\name{plot_signals}
\alias{plot_signals}
\title{Plot signals line by line}
\usage{
plot_signals(
  signals,
  sample_rate = 1,
  col = 1,
  space = 0.995,
  space_mode = "quantile",
  start_time = 0,
  duration = NULL,
  compress = TRUE,
  channel_names = NULL,
  ylab = "Channel",
  time_shift = 0,
  lwd = 0.5,
  cex = 2,
  new_plot = T,
  plot = "base",
  xlim = NULL,
  ...
)
}
\arguments{
\item{signals}{signals to plot, with each row one signal}

\item{sample_rate}{sample rate}

\item{col}{Color, either length of 1 or number of signals. Can be numeric or color name}

\item{space}{space between to signals. If \code{space_mode='quantile'}, then
space is determined by quantile of signal (0 - 1). If \code{space_mode='absoute'}, then
space will be as is.}

\item{start_time}{Time in seconds at which time point the signal should be drawn}

\item{duration}{length of 1. Time in seconds the duration of time to be drawn.
Default is NULL (Total time range)}

\item{compress}{FALSE means no compression for signals, TRUE is auto-detection,
2, 3, 4,... means compress signals by x and then plot. (usually compress signal to save time)}

\item{channel_names}{Names for each signals. Will be Y tick labels}

\item{ylab}{Y axis label}

\item{plot, xlim, space_mode, time_shift, lwd, cex, new_plot}{Deprecated}

\item{...}{pass to \code{\link[graphics]{matplot}}}
}
\description{
Plot signals line by line
}
