% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-print.R
\name{print.rave_printable}
\alias{print.rave_printable}
\title{Override print method for printable environment}
\usage{
\method{print}{rave_printable}(x, ...)
}
\arguments{
\item{x}{an rave_printable environment}

\item{...}{passed from or to other methods}
}
\description{
Override print method for printable environment
}
