% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress.R
\name{progress}
\alias{progress}
\title{A wrapper for shiny Progress object}
\usage{
progress(
  title,
  max = 1,
  session = getDefaultReactiveDomain(),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{title}{Main title for progress bar}

\item{max}{How many steps you have for this process}

\item{session}{shiny session, default is \code{\link[shiny]{getDefaultReactiveDomain}}}

\item{quiet}{nonreactive-only mode? default is FALSE. If TRUE, then progress bar will be hidden in shiny app}

\item{...}{other parameters passing to \code{\link[dipsaus]{progress2}}}
}
\description{
A wrapper for shiny Progress object
}
\details{
shiny::Progress class cannot be used under non-reactive environment.
rave::progress function wrap it up so that you can use it in non-reactive settings.
}
