% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-pwelch.R
\name{pwelch}
\alias{pwelch}
\title{Plot "Welch" Periodogram}
\usage{
pwelch(
  x,
  fs,
  window = 64,
  noverlap = 8,
  nfft = 256,
  col = "black",
  xlim = NULL,
  ylim = NULL,
  main = "Welch periodogram",
  plot = TRUE,
  log = "xy",
  spec_func = stats::spectrum,
  cex = 1,
  ...
)
}
\arguments{
\item{x}{signal}

\item{fs}{sample rate}

\item{window}{window length, default 128}

\item{noverlap}{overlap between two adjacent windows, by default is 8}

\item{nfft}{number of basis functions}

\item{col, xlim, ylim, main, cex, ...}{will be passed to plot}

\item{plot}{logical, plot the result or not}

\item{log}{indicates which axis should be \code{\link{log10}} value. options 
are \code{''}, \code{'x'}, \code{'y'}, \code{'xy'}.}

\item{spec_func}{deprecated}
}
\description{
Plot "Welch" Periodogram
}
