% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Subject.R
\name{r_to_py.Subject}
\alias{r_to_py.Subject}
\title{Convert subject to python object}
\usage{
r_to_py.Subject(obj, convert = FALSE)
}
\arguments{
\item{obj}{Subject class}

\item{convert}{pass to \code{r_to_py} in \code{reticulate} package}
}
\description{
Convert subject to python object
}
