% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_brain2.R
\name{rave_brain2}
\alias{rave_brain2}
\title{Tools to load and view brain in 3D viewer}
\usage{
rave_brain2(surfaces = "pial", multiple_subject = FALSE,
  prefix = "std.141.")
}
\arguments{
\item{surfaces}{one or more from `pial`, `white`, `smoothwm`, brain surface types}

\item{multiple_subject}{is this a template brain?}

\item{prefix}{internally used, prefix to `FreeSurfer` asc files}
}
\description{
Tools to load and view brain in 3D viewer
}
