% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-options.R
\name{rave_options}
\alias{rave_options}
\title{Function to change rave-options}
\usage{
rave_options(
  ...,
  .save = TRUE,
  launch_gui = TRUE,
  host = "127.0.0.1",
  port = NULL
)
}
\arguments{
\item{...}{Key-Value option pairs}

\item{.save}{save to disk? ignored most of the time}

\item{launch_gui}{launch shiny app?}

\item{host}{IP address of host}

\item{port}{Port number}
}
\description{
Function to change rave-options
}
