% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-preprocess.R
\name{rave_preprocess}
\alias{rave_preprocess}
\title{RAVE Preprocess Function}
\usage{
rave_preprocess(sidebar_width = 3, launch.browser = T,
  host = "127.0.0.1", port = NULL, quiet = T, test.mode = F,
  modules, ver = "3", theme = "purple", ...)
}
\arguments{
\item{sidebar_width}{sibar width from 1 to 12 recommended 3 or 4}

\item{launch.browser}{launch browser default is on}

\item{host}{default is localhost 127.0.0.1}

\item{port}{integer port if the app}

\item{quiet}{soft deprecated}

\item{test.mode}{passed to shinyApp}

\item{modules}{preprocess modules to load, reserved for future use}

\item{ver}{internally used please don't change}

\item{theme}{color theme}

\item{...}{not used}
}
\description{
RAVE Preprocess Function
}
