% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{rave_preprocess}
\alias{rave_preprocess}
\title{RAVE Preprocess Function}
\usage{
rave_pre_process(sidebar_width = 3L, launch.browser = T, host = '127.0.0.1', ...)
}
\description{
RAVE Preprocess Function
}
