% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-preprocess.R
\name{rave_preprocess}
\alias{rave_preprocess}
\title{RAVE Preprocess Function}
\usage{
rave_preprocess(
  sidebar_width = 3,
  launch.browser = TRUE,
  host = "127.0.0.1",
  port = NULL,
  quiet = TRUE,
  beta = FALSE,
  test.mode = FALSE,
  modules,
  ver = "3",
  theme = "purple",
  ...
)
}
\arguments{
\item{sidebar_width}{sidebar width from 1 to 11.}

\item{launch.browser}{whether to launch browser, default is on}

\item{host}{default is \code{"localhost"}}

\item{port}{integer port of the app}

\item{quiet}{soft deprecated}

\item{beta}{whether to load experimental modules, default is false}

\item{test.mode}{passed to \code{\link[shiny]{shinyApp}}}

\item{modules}{preprocess modules to load, reserved}

\item{ver}{internally used please don't change}

\item{theme}{color theme}

\item{...}{used for other functions for configuration and debug only}
}
\description{
RAVE Preprocess Function
}
