% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_history.R
\docType{data}
\name{RAVEHistory}
\alias{RAVEHistory}
\title{R6 class for small object caching}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RAVEHistory
}
\description{
R6 class for small object caching
}
\keyword{datasets}
