% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_io.R
\name{read_mgrid}
\alias{read_mgrid}
\title{Make SUMA grid (I do not know what's this, John wrote it)}
\usage{
read_mgrid(con, raw = F)
}
\arguments{
\item{con}{file}

\item{raw}{raw file or processed I guess}
}
\description{
Make SUMA grid (I do not know what's this, John wrote it)
}
\author{
J.Magnnotti
}
