% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-afnisuma.R
\name{read_mgrid}
\alias{read_mgrid}
\title{Make \code{iElvis} \code{mgrid} file}
\usage{
read_mgrid(con, raw = F)
}
\arguments{
\item{con}{\code{mgrid} file}

\item{raw}{raw file or processed I guess}
}
\description{
Make \code{iElvis} \code{mgrid} file
}
\author{
John Magnotti
}
