% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{LazyFST}
\alias{load_fst_or_h5}
\alias{LazyH5}
\alias{load_h5}
\alias{save_h5}
\alias{read_mat}
\alias{Tensor}
\alias{ECoGTensor}
\alias{catgl}
\alias{clear_env}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dipsaus}{\code{\link[dipsaus]{clear_env}}}

  \item{raveio}{\code{\link[raveio]{ECoGTensor}}, \code{\link[raveio]{LazyFST}}, \code{\link[raveio]{LazyH5}}, \code{\link[raveio]{Tensor}}, \code{\link[raveio]{catgl}}, \code{\link[raveio]{load_fst_or_h5}}, \code{\link[raveio]{load_h5}}, \code{\link[raveio]{read_mat}}, \code{\link[raveio]{save_h5}}}
}}

