% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-csv.R
\name{safe_write_csv}
\alias{safe_write_csv}
\title{Save data to "CSV", if file exists, rename old file}
\usage{
safe_write_csv(data, file, ..., quiet = FALSE)
}
\arguments{
\item{data}{data frame}

\item{file}{"CSV" file to save}

\item{...}{pass to \code{\link[utils]{write.csv}}}

\item{quiet}{suppress overwrite message}
}
\description{
Save data to "CSV", if file exists, rename old file
}
