% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-hdf5.R
\name{save_h5}
\alias{save_h5}
\title{Save objects to "HDF5" file without trivial checks}
\usage{
save_h5(
  x,
  file,
  name,
  chunk = "auto",
  level = 4,
  replace = TRUE,
  new_file = FALSE,
  ctype = NULL,
  ...
)
}
\arguments{
\item{x}{array, matrix, or vector}

\item{file}{\code{HDF5} file}

\item{name}{path to dataset in format like \code{"group/data_name"}}

\item{chunk}{chunk size}

\item{level}{compress level}

\item{replace}{if dataset exists, replace?}

\item{new_file}{remove old file if exists?}

\item{ctype}{dataset type: numeric? character?}

\item{...}{passed to other \code{LazyH5$save}}
}
\description{
Save objects to "HDF5" file without trivial checks
}
\examples{

file <- tempfile()
x <- 1:120; dim(x) <- 2:5

# save x to file with name /group/dataset/1
save_h5(x, file, '/group/dataset/1', chunk = dim(x))

# read data
y <- load_h5(file, '/group/dataset/1')
y[]
}
\seealso{
\code{\link{load_h5}}
}
