% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io_h5.R
\name{save_h5}
\alias{save_h5}
\title{Save objects to H5 file without trivial checkings}
\usage{
save_h5(x, file, name, chunk = "auto", level = 4, replace = TRUE,
  new_file = FALSE, ctype = NULL, ...)
}
\description{
Save objects to H5 file without trivial checkings
}
\examples{
\dontrun{
file <- tempfile()
x <- 1:120; dim(x) <- 2:5

# save x to file with name /group/dataset/1
save_h5(x, file, '/group/dataset/1', chunk = dim(x))

# read data
y <- load_h5(file, '/group/dataset/1')
y[]
}
}
\seealso{
\code{\link{load_h5}}
}
