% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-shinirize.R
\name{shinirize}
\alias{shinirize}
\title{Convert module to objects used in shiny}
\usage{
shinirize(module, session = getDefaultReactiveDomain(), test.mode = TRUE)
}
\arguments{
\item{module}{ModuleEnvir object}

\item{session}{shiny session, default let shiny decide}

\item{test.mode}{passed by init_app}
}
\description{
Convert module to objects used in shiny
}
