% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-main_app.R
\name{start_rave}
\alias{start_rave}
\alias{start_rave_legacy}
\alias{launch_demo}
\alias{start_rave2}
\title{Start RAVE main application}
\usage{
start_rave_legacy(
  modules = NULL,
  active_module = NULL,
  launch.browser = TRUE,
  theme = "purple",
  disable_sidebar = FALSE,
  simplify_header = FALSE,
  token = NULL,
  data_repo = getDefaultDataRepository(),
  ...
)

launch_demo(
  modules = "power_explorer",
  launch.browser = TRUE,
  theme = "green",
  disable_sidebar = TRUE,
  simplify_header = FALSE,
  ...
)

start_rave2(
  host = "127.0.0.1",
  port = NULL,
  launch.browser = TRUE,
  jupyter = FALSE,
  as_job = FALSE,
  ...
)

start_rave(
  host = "127.0.0.1",
  port = NULL,
  launch.browser = TRUE,
  jupyter = FALSE,
  as_job = FALSE,
  ...
)
}
\arguments{
\item{modules}{character vector, modules modules to load before starting application.}

\item{active_module}{character, which module to show as default.}

\item{launch.browser}{logical, whether to launch browser.}

\item{theme}{character, color theme, default is \code{'purple'}.}

\item{disable_sidebar}{logical, whether to hide sidebar.}

\item{simplify_header}{logical, whether to show simplified header.}

\item{token}{character vector, default is \code{NULL}. If specified, then
a \code{?token=...} is needed in url to access to the application.}

\item{data_repo}{internally used.}

\item{...}{other parameters. See details.}

\item{host, port, jupyter, as_job}{'RAVE' 2.0 related arguments; see 
\code{\link[ravedash]{start_session}}}
}
\description{
Start RAVE main application
}
