% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-main_app.R
\name{start_rave}
\alias{start_rave}
\title{Start RAVE main application}
\usage{
start_rave(
  modules = NULL,
  active_module = NULL,
  launch.browser = TRUE,
  theme = "purple",
  disable_sidebar = FALSE,
  simplify_header = FALSE,
  token = NULL,
  data_repo = getDefaultDataRepository(),
  ...
)
}
\arguments{
\item{modules}{character vector, modules modules to load before starting application.}

\item{active_module}{character, which module to show as default.}

\item{launch.browser}{logical, whether to launch browser.}

\item{theme}{character, color theme, default is \code{'purple'}.}

\item{disable_sidebar}{logical, whether to hide sidebar.}

\item{simplify_header}{logical, whether to show simplified header.}

\item{token}{character vector, default is \code{NULL}. If specified, then
a \code{?token=...} is needed in url to access to the application.}

\item{data_repo}{internally used.}

\item{...}{other parameters. See details.}
}
\description{
Start RAVE main application
}
