% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI-main_app.R
\name{start_yael}
\alias{start_yael}
\title{Start 'YAEL' electrode localization}
\usage{
start_yael(
  host = "127.0.0.1",
  port = NULL,
  launch.browser = TRUE,
  as_job = FALSE,
  ...
)
}
\arguments{
\item{host}{host IP address}

\item{port}{integer port number; default is random}

\item{launch.browser}{whether to launch browsers}

\item{as_job}{whether to launch in background; available only in 'RStudio'}

\item{...}{passed to \code{\link[ravedash]{start_session}}}
}
\description{
Start 'YAEL' electrode localization
}
