% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-subject.R
\name{subject_tmpfile}
\alias{subject_tmpfile}
\title{Create temp file in subject module folder}
\usage{
subject_tmpfile(
  module_id,
  fun_name = "",
  project_name,
  subject_code,
  pattern = "file_",
  data_env = getDefaultDataRepository()
)
}
\arguments{
\item{module_id}{module id}

\item{fun_name}{function name (usually export_"function_name" in the module)}

\item{project_name}{project name}

\item{subject_code}{subject code}

\item{pattern}{passed to \code{\link{tempfile}}}

\item{data_env}{internally used}
}
\description{
Create temp file in subject module folder
}
