% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_io.R
\name{suma_spec_parse}
\alias{suma_spec_parse}
\title{Parse spec file}
\usage{
suma_spec_parse(subject)
}
\arguments{
\item{subject}{Either characters such as 'Project/Subject' or Subject object created by Subject$new(...)}
}
\description{
Parse spec file
}
\examples{
\dontrun{
subject = 'Demo/Subject'
# or create subject object
subject = Subject$new('Demo', 'Subject')

suma_spec_parse(subject)
}
}
