% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-afnisuma.R
\name{suma_spec_parse}
\alias{suma_spec_parse}
\title{Parse \code{SUMA} \code{spec} file}
\usage{
suma_spec_parse(subject, spec_file)
}
\arguments{
\item{subject}{either characters in format like \code{'Project/Subject'} or 
\code{\link[rave]{Subject}} object created by \code{Subject$new(...)}}

\item{spec_file}{default decided by \code{rave_options('suma_spec_file')}, 
depending on subjects}
}
\description{
Parse \code{SUMA} \code{spec} file
}
\examples{

\dontrun{
subject = 'Demo/YAB'
# or create subject object
subject = Subject$new('Demo', 'YAB')

# Please download sample subjects first to run
suma_spec_parse(subject)
}
}
