% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_io.R
\name{suma_surface_volume_parse}
\alias{suma_surface_volume_parse}
\title{Parse surface volumn file header}
\usage{
suma_surface_volume_parse(file_path)
}
\arguments{
\item{file_path}{path to brik or head file}
}
\description{
Parse surface volumn file header
}
