% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-afnisuma.R
\name{suma_surface_volume_parse}
\alias{suma_surface_volume_parse}
\title{Parse \code{AFNI} \code{BRIK/HEAD} file}
\usage{
suma_surface_volume_parse(file_path)
}
\arguments{
\item{file_path}{path to \code{BRIK} or \code{HEAD} file}
}
\description{
Parse \code{AFNI} \code{BRIK/HEAD} file
}
