% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Tensor.R
\name{Tensor}
\alias{Tensor}
\title{Wrapper class for tensor arrays}
\description{
Wrapper class for tensor arrays

Wrapper class for tensor arrays
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-finalize}{\code{Tensor$finalize()}}
\item \href{#method-print}{\code{Tensor$print()}}
\item \href{#method-.use_multi_files}{\code{Tensor$.use_multi_files()}}
\item \href{#method-new}{\code{Tensor$new()}}
\item \href{#method-subset}{\code{Tensor$subset()}}
\item \href{#method-flatten}{\code{Tensor$flatten()}}
\item \href{#method-to_swap}{\code{Tensor$to_swap()}}
\item \href{#method-to_swap_now}{\code{Tensor$to_swap_now()}}
\item \href{#method-get_data}{\code{Tensor$get_data()}}
\item \href{#method-set_data}{\code{Tensor$set_data()}}
\item \href{#method-collapse}{\code{Tensor$collapse()}}
\item \href{#method-operate}{\code{Tensor$operate()}}
\item \href{#method-clone}{\code{Tensor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.use_multi_files"></a>}}
\subsection{Method \code{.use_multi_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$.use_multi_files(mult)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$new(
  data,
  dim,
  dimnames,
  varnames,
  hybrid = F,
  use_index = F,
  swap_file = tempfile(),
  temporary = TRUE,
  multi_files = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$subset(..., drop = FALSE, data_only = F, .env = parent.frame())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-flatten"></a>}}
\subsection{Method \code{flatten()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$flatten(include_index = F, value_name = "value")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_swap"></a>}}
\subsection{Method \code{to_swap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$to_swap(use_index = F, delay = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_swap_now"></a>}}
\subsection{Method \code{to_swap_now()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$to_swap_now(use_index = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$get_data(drop = F, gc_delay = 3)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$set_data(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collapse"></a>}}
\subsection{Method \code{collapse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$collapse(keep, method = "mean")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-operate"></a>}}
\subsection{Method \code{operate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$operate(
  by,
  fun = .Primitive("/"),
  match_dim,
  mem_optimize = F,
  same_dimension = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tensor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
