% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_tensor.R
\docType{data}
\name{Tensor}
\alias{Tensor}
\title{Wrapper class for tensor arrays}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tensor
}
\description{
Wrapper class for tensor arrays
}
\keyword{datasets}
