% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_rave_options.R
\name{test_hdspeed}
\alias{test_hdspeed}
\title{Function to test local disk speed}
\usage{
test_hdspeed(file_size = 1e+07, quiet = F)
}
\arguments{
\item{file_size}{default 10MB, i.e. 1e7}

\item{quiet}{show messages?}
}
\description{
Function to test local disk speed
}
