% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-wavelet.R
\name{wavelet}
\alias{wavelet}
\title{Wavelet Transformation With Phase}
\usage{
wavelet(data, freqs, srate, wave_num, demean = TRUE)
}
\arguments{
\item{data}{- vector of time series to be decomposed}

\item{freqs}{- vector of center frequencies for decomposition}

\item{srate}{- sample rate (in Hz)}

\item{wave_num}{- desired number of cycles in wavelet (typically 3-20 for 
frequencies 2-200).}

\item{demean}{- whether to remove the mean of data first?}
}
\description{
The code was translated from Matlab script written by 
Brett Foster, Stanford Memory Lab, 2015 with permission to use in `RAVE`.
}
\details{
Decompose time series data into time-frequency
representation (spectral decomposition) using wavelet transform. Employs
"Morlet" wavelet method (gaussian taper sine wave) to obtain the analytic
signal for specified frequencies (via convolution).
}
